% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-BindMatrix.R
\name{BPCells-bind}
\alias{BPCells-bind}
\alias{rbind2}
\alias{cbind2}
\alias{rbind}
\alias{cbind}
\alias{arbind}
\alias{acbind}
\alias{bindROWS}
\alias{bindCOLS}
\alias{rbind2,BPCellsMatrix,BPCellsMatrix-method}
\alias{rbind,BPCellsMatrix-method}
\alias{arbind,BPCellsMatrix-method}
\alias{bindROWS,BPCellsMatrix-method}
\alias{cbind2,BPCellsMatrix,BPCellsMatrix-method}
\alias{cbind,BPCellsMatrix-method}
\alias{acbind,BPCellsMatrix-method}
\alias{bindCOLS,BPCellsMatrix-method}
\title{Combine two Objects by Columns or Rows}
\usage{
\S4method{rbind2}{BPCellsMatrix,BPCellsMatrix}(x, y, mode = NULL, ..., threads = 0L)

\S4method{rbind}{BPCellsMatrix}(
  ...,
  mode = NULL,
  threads = 0L,
  use.first.dimnames = TRUE,
  deparse.level = 1L
)

\S4method{arbind}{BPCellsMatrix}(..., mode = NULL, threads = 0L)

\S4method{bindROWS}{BPCellsMatrix}(
  x,
  objects = list(),
  use.names = TRUE,
  ignore.mcols = TRUE,
  check = TRUE
)

\S4method{cbind2}{BPCellsMatrix,BPCellsMatrix}(x, y, mode = NULL, ..., threads = 0L)

\S4method{cbind}{BPCellsMatrix}(
  ...,
  mode = NULL,
  threads = 0L,
  use.first.dimnames = TRUE,
  deparse.level = 1L
)

\S4method{acbind}{BPCellsMatrix}(..., mode = NULL, threads = 0L, use.first.dimnames = TRUE)

\S4method{bindCOLS}{BPCellsMatrix}(
  x,
  objects = list(),
  use.names = TRUE,
  ignore.mcols = TRUE,
  check = TRUE
)
}
\arguments{
\item{x, y}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object or \link[=BPCellsDelayedOp-class]{BPCellsSeed} object.}

\item{mode}{Storage mode of BPCells matrix, one of \code{uint32_t} (unsigned
32-bit integer), \code{float} (32-bit real number), or \code{double} (64-bit real
number). R cannot differentiate 32-bit and 64-bit real number, so
\link[=BPCellsSeed-class]{type} method always return "double" for both \code{float}
and \code{double} mode.}

\item{...}{\itemize{
\item \code{rbind2} and \code{cbind2}: Not used currently.
\item \code{rbind}, \code{arbind}, \code{cbind}, and \code{acbind}: A list of \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}}

\item{threads}{Set the number of threads to use for sparse-dense multiply and
\link[BPCells:matrix_stats]{matrix_stats}.}

\item{use.first.dimnames}{Ignored, always be \code{TRUE} in BPCells.}

\item{deparse.level}{Ignored, used by generic methods.}

\item{objects}{
    A list of S4 objects to bind to \code{x}. They should typically (but not
    necessarily) have the same class as \code{x}.
  }

\item{use.names}{Ignored, always be \code{TRUE}.}

\item{ignore.mcols}{Ignored.}

\item{check}{Ignored.}
}
\value{
If \code{mode} is specified, the mode of all specified object will be converted.
\itemize{
\item \code{cbind2}, \code{acbind}, \code{cbind}, \code{bindCOLS}: A \link[=BPCellsMatrix-class]{BPCellsMatrix} object combined by
columns.
\item \code{rbind2}, \code{arbind}, \code{rbind}, \code{bindROWS}: A \link[=BPCellsMatrix-class]{BPCellsMatrix} object combined by
rows.
}
}
\description{
Combine two Objects by Columns or Rows
}
\seealso{
\link{convert_mode}
}
