% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-Compare.R
\name{BPCells-Compare}
\alias{BPCells-Compare}
\alias{binarize}
\alias{binarize,BPCellsMatrix-method}
\alias{<,numeric,BPCellsMatrix-method}
\alias{>,BPCellsMatrix,numeric-method}
\alias{<=,numeric,BPCellsMatrix-method}
\alias{>=,BPCellsMatrix,numeric-method}
\title{Convert matrix elements to zeros and ones}
\usage{
binarize(object, ...)

\S4method{binarize}{BPCellsMatrix}(object, ...)

\S4method{<}{numeric,BPCellsMatrix}(e1, e2)

\S4method{>}{BPCellsMatrix,numeric}(e1, e2)

\S4method{<=}{numeric,BPCellsMatrix}(e1, e2)

\S4method{>=}{BPCellsMatrix,numeric}(e1, e2)
}
\arguments{
\item{object}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{...}{
  Arguments passed on to \code{\link[BPCells:binarize]{BPCells::binarize}}
  \describe{
    \item{\code{threshold}}{A numeric value that determines whether the elements
of x are set to zero or one.}
    \item{\code{strict_inequality}}{A logical value determining whether the
comparison to the threshold is >= (strict_inequality=FALSE)
or > (strict_inequality=TRUE).}
  }}

\item{e1, e2}{One of \code{x} or \code{y} must be a \link[=BPCellsMatrix-class]{BPCellsMatrix} object, and the another can
be a \link[=BPCellsMatrix-class]{BPCellsMatrix} object or a matrix-like object which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix} object.}
}
\value{
A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
Binarize compares the matrix element values to the
threshold value and sets the output elements to either zero
or one. By default, element values greater than the threshold
are set to one; otherwise, set to zero. When strict_inequality
is set to FALSE, element values greater than or equal to the
threshold are set to one. As an alternative, the \code{<}, \code{<=}, \code{>},
and \code{>=} operators are also supported.
}
\note{
Methods listed here are supported by \code{BPCells}, other
\link[methods:S4groupGeneric]{Compare} operators will use the methods defined in
\link[=DelayedArray-utils]{DelayedArray}.
}
