% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-crossprod.R
\name{BPCells-crossprod}
\alias{BPCells-crossprod}
\alias{crossprod}
\alias{crossprod,BPCellsMatrix,BPCellsMatrix-method}
\alias{crossprod,BPCellsMatrix,ANY-method}
\alias{crossprod,ANY,BPCellsMatrix-method}
\alias{crossprod,BPCellsMatrix,matrix-method}
\alias{crossprod,matrix,BPCellsMatrix-method}
\alias{crossprod,BPCellsMatrix,numeric-method}
\alias{crossprod,numeric,BPCellsMatrix-method}
\title{Matrix Crossproduct}
\usage{
\S4method{crossprod}{BPCellsMatrix,BPCellsMatrix}(x, y)

\S4method{crossprod}{BPCellsMatrix,ANY}(x, y)

\S4method{crossprod}{ANY,BPCellsMatrix}(x, y)

\S4method{crossprod}{BPCellsMatrix,matrix}(x, y)

\S4method{crossprod}{matrix,BPCellsMatrix}(x, y)

\S4method{crossprod}{BPCellsMatrix,numeric}(x, y)

\S4method{crossprod}{numeric,BPCellsMatrix}(x, y)
}
\arguments{
\item{x, y}{One of \code{x} or \code{y} must be a \link[=BPCellsMatrix-class]{BPCellsMatrix} object, and the another must
be a \link[=BPCellsMatrix-class]{BPCellsMatrix} object or a matrix-like object which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix} object.}
}
\value{
A dense matrix if one of \code{x} or \code{y} is a regular matrix or atomic
vector. Otherwise, a \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
Given matrices x and y as arguments, return a matrix cross-product.
}
\seealso{
\itemize{
\item \link[=BPCells-Multiplication]{\%*\%}
\item \link[=BPCells-tcrossprod]{tcrossprod}
}
}
