% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-Math.R
\name{BPCells-Math}
\alias{BPCells-Math}
\alias{expm1_slow}
\alias{expm1_slow,BPCellsMatrix-method}
\alias{expm1,BPCellsMatrix-method}
\alias{expm1}
\alias{log1p,BPCellsMatrix-method}
\alias{log1p}
\alias{log1p_single}
\alias{log1p_single,BPCellsMatrix-method}
\alias{round,BPCellsMatrix-method}
\alias{round}
\title{Math operators for BPCellsMatrix}
\usage{
expm1_slow(x)

\S4method{expm1_slow}{BPCellsMatrix}(x)

\S4method{expm1}{BPCellsMatrix}(x)

\S4method{log1p}{BPCellsMatrix}(x)

log1p_single(x)

\S4method{log1p_single}{BPCellsMatrix}(x)

\S4method{round}{BPCellsMatrix}(x, digits = 0)
}
\arguments{
\item{x}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{digits}{Integer indicating the number of decimal places}
}
\value{
\itemize{
\item \code{expm1} and \code{expm1_slow}: compute \code{exp(x)-1} of matrix.
}

\itemize{
\item \code{log1p} and \code{log1p_single}: compute \code{log(1+x)} of matrix. \code{log1p_single}
use single-precision math in intermediate stages, which is 2x faster than
\code{log1p}.
}

\itemize{
\item \code{round}: Rounding of matrix Numbers.
}
}
\description{
Math operators for BPCellsMatrix
}
\note{
Methods listed here are supported by \code{BPCells}, other
\link[methods:S4groupGeneric]{Math} or \link[methods:S4groupGeneric]{Math2} operators will use the
methods defined in \link[=DelayedArray-utils]{DelayedArray}.
}
