% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Multiply.R
\name{BPCells-Multiplication}
\alias{BPCells-Multiplication}
\alias{\%*\%,BPCellsMatrix,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,dgCMatrix-method}
\alias{\%*\%,dgCMatrix,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,ANY-method}
\alias{\%*\%,ANY,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,matrix-method}
\alias{\%*\%,matrix,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,numeric-method}
\alias{\%*\%,numeric,BPCellsMatrix-method}
\alias{\%*\%,BPCellsSeed,BPCellsSeed-method}
\alias{\%*\%,BPCellsSeed,dgCMatrix-method}
\alias{\%*\%,dgCMatrix,BPCellsSeed-method}
\alias{\%*\%,BPCellsSeed,matrix-method}
\alias{\%*\%,matrix,BPCellsSeed-method}
\alias{\%*\%,BPCellsSeed,numeric-method}
\alias{\%*\%,numeric,BPCellsSeed-method}
\title{Matrix Multiplication}
\usage{
\S4method{\%*\%}{BPCellsMatrix,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,dgCMatrix}(x, y)

\S4method{\%*\%}{dgCMatrix,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,ANY}(x, y)

\S4method{\%*\%}{ANY,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,matrix}(x, y)

\S4method{\%*\%}{matrix,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,numeric}(x, y)

\S4method{\%*\%}{numeric,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsSeed,BPCellsSeed}(x, y)

\S4method{\%*\%}{BPCellsSeed,dgCMatrix}(x, y)

\S4method{\%*\%}{dgCMatrix,BPCellsSeed}(x, y)

\S4method{\%*\%}{BPCellsSeed,matrix}(x, y)

\S4method{\%*\%}{matrix,BPCellsSeed}(x, y)

\S4method{\%*\%}{BPCellsSeed,numeric}(x, y)

\S4method{\%*\%}{numeric,BPCellsSeed}(x, y)
}
\arguments{
\item{x, y}{One of \code{x} or \code{y} must be \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object, and the another must be a matrix
which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix}.}
}
\value{
Matrix multiplication, a \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object depends on the class of \code{x} or
\code{y}.
}
\description{
Multiplies two matrices, if they are conformable. If one argument is a
vector, it will be promoted to either a row or column matrix to make the two
arguments conformable. If both are vectors of the same length, it will return
the inner product (as a matrix).
}
