% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-summarization.R
\name{BPCells-Summarization}
\alias{BPCells-Summarization}
\alias{rowSums}
\alias{rowSums,BPCellsMatrix-method}
\alias{colSums,BPCellsMatrix-method}
\alias{colSums}
\alias{rowMeans,BPCellsMatrix-method}
\alias{rowMeans}
\alias{colMeans,BPCellsMatrix-method}
\alias{colMeans}
\alias{rowVars,BPCellsMatrix-method}
\alias{rowVars}
\alias{colVars,BPCellsMatrix-method}
\alias{colVars}
\title{BPCellsMatrix row/col summarization}
\usage{
\S4method{rowSums}{BPCellsMatrix}(x)

\S4method{colSums}{BPCellsMatrix}(x)

\S4method{rowMeans}{BPCellsMatrix}(x)

\S4method{colMeans}{BPCellsMatrix}(x)

\S4method{rowVars}{BPCellsMatrix}(x)

\S4method{colVars}{BPCellsMatrix}(x)
}
\arguments{
\item{x}{A \link{BPCellsMatrix} object.}
}
\value{
\itemize{
\item \code{rowSums()}: vector of row sums
}

\itemize{
\item \code{colSums()}: vector of col sums
}

\itemize{
\item \code{rowMeans()}: vector of row means
}

\itemize{
\item \code{colMeans()}: vector of col means
}

\itemize{
\item \code{rowVars()}: vector of row vars
}

\itemize{
\item \code{colVars()}: vector of col vars
}
}
\description{
BPCellsMatrix row/col summarization
}
