% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-tcrossprod.R
\name{BPCells-tcrossprod}
\alias{BPCells-tcrossprod}
\alias{tcrossprod}
\alias{tcrossprod,BPCellsMatrix,BPCellsMatrix-method}
\alias{tcrossprod,BPCellsMatrix,ANY-method}
\alias{tcrossprod,ANY,BPCellsMatrix-method}
\alias{tcrossprod,BPCellsMatrix,matrix-method}
\alias{tcrossprod,matrix,BPCellsMatrix-method}
\alias{tcrossprod,BPCellsMatrix,numeric-method}
\alias{tcrossprod,numeric,BPCellsMatrix-method}
\title{Matrix Products of Transpose}
\usage{
\S4method{tcrossprod}{BPCellsMatrix,BPCellsMatrix}(x, y)

\S4method{tcrossprod}{BPCellsMatrix,ANY}(x, y)

\S4method{tcrossprod}{ANY,BPCellsMatrix}(x, y)

\S4method{tcrossprod}{BPCellsMatrix,matrix}(x, y)

\S4method{tcrossprod}{matrix,BPCellsMatrix}(x, y)

\S4method{tcrossprod}{BPCellsMatrix,numeric}(x, y)

\S4method{tcrossprod}{numeric,BPCellsMatrix}(x, y)
}
\arguments{
\item{x, y}{One of \code{x} or \code{y} must be a \link[=BPCellsMatrix-class]{BPCellsMatrix} object, and the another must
be a \link[=BPCellsMatrix-class]{BPCellsMatrix} object or a matrix-like object which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix} object.}
}
\value{
A dense matrix if one of \code{x} or \code{y} is a regular matrix or atomic
vector. Otherwise, a \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
Given matrices x and y as arguments, return a matrix product of Transpose.
}
\seealso{
\itemize{
\item \link[=BPCells-Multiplication]{\%*\%}
\item \link[=BPCells-crossprod]{crossprod}
}
}
