% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO-10xHDF5.R
\name{BPCells10xHDF5-IO}
\alias{BPCells10xHDF5-IO}
\alias{readBPCells10xHDF5Matrix}
\alias{writeBPCells10xHDF5Matrix}
\alias{writeBPCells10xHDF5Matrix,ANY-method}
\title{Read/write sparse matrices from (or into) 10x feature matrix}
\usage{
readBPCells10xHDF5Matrix(
  path,
  feature_type = NULL,
  buffer_size = 16384L,
  seedform = NULL
)

writeBPCells10xHDF5Matrix(x, ...)

\S4method{writeBPCells10xHDF5Matrix}{ANY}(
  x,
  path,
  barcodes = colnames(x),
  feature_ids = rownames(x),
  feature_names = rownames(x),
  feature_types = "Gene Expression",
  feature_metadata = list(),
  buffer_size = 16384L,
  chunk_size = 1024L,
  gzip = 0L,
  seedform = NULL
)
}
\arguments{
\item{path}{A string path of the \verb{10x HDF5} file to read or save data into.}

\item{feature_type}{Optional selection of feature types to include in output matrix.
For multiome data, the options are "Gene Expression" and "Peaks". This option is
only compatible with files from cellranger 3.0 and newer.}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}

\item{seedform}{A string, \code{"BPCells"} or \code{"DelayedArray"}. If \code{NULL}, will
use the the default \code{seedform}. Details see \link{seedform}.}

\item{x}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object or a matrix-like object which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix} object.}

\item{...}{Additional arguments passed into specific methods.}

\item{barcodes}{Vector of names for the cells}

\item{feature_ids}{Vector of IDs for the features}

\item{feature_names}{Vector of names for the features}

\item{feature_types}{String or vector of feature types}

\item{feature_metadata}{Named list of additional metadata vectors
to store for each feature}

\item{chunk_size}{For performance tuning only. The chunk size used for the HDF5 array storage.}

\item{gzip}{Gzip compression level. Default is 0 (no gzip compression).}
}
\value{
A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
\itemize{
\item \code{readBPCells10xHDF5Matrix}: read a sparce matrices from a HDF5 file on disk
\item \code{writeBPCells10xHDF5Matrix}: Write a sparce matrices into a HDF5 file on
disk
}
}
\details{
The 10x format makes use of gzip compression for the matrix data,
which can slow down read performance. Consider writing into another format
if the read performance is important to you.

Input matrices must be in column-major storage order,
and if the rownames and colnames are not set, names must be
provided for the relevant metadata parameters. Some of the
metadata parameters are not read by default in BPCells, but
it is possible to export them for use with other tools.
}
\seealso{
\itemize{
\item \link{BPCellsSeed}
\item \link{BPCellsDir-IO}
\item \link{BPCellsHDF5-IO}
\item \link{BPCellsMem-IO}
\item \link{BPCells10xHDF5-IO}
}
}
