% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO-AnnHDF5.R
\name{BPCellsAnnHDF5-IO}
\alias{BPCellsAnnHDF5-IO}
\alias{readBPCellsAnnHDF5Matrix}
\alias{writeBPCellsAnnHDF5Matrix}
\alias{writeBPCellsAnnHDF5Matrix,ANY-method}
\title{Read/write sparse matrices from (or into) AnnData HDF5 file}
\usage{
readBPCellsAnnHDF5Matrix(path, group = "X", buffer_size = 16384L)

writeBPCellsAnnHDF5Matrix(x, ...)

\S4method{writeBPCellsAnnHDF5Matrix}{ANY}(
  x,
  path,
  group = "X",
  buffer_size = 16384L,
  chunk_size = 1024L,
  gzip = 0L
)
}
\arguments{
\item{path}{A string path of the AnnData HDF5 file to read or save data
into.}

\item{group}{The group within the hdf5 file to write the data to. If writing
to an existing hdf5 file this group must not already be in use}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}

\item{x}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object or a matrix-like object which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix} object.}

\item{...}{Additional arguments passed into specific methods.}

\item{chunk_size}{For performance tuning only. The chunk size used for the HDF5 array storage.}

\item{gzip}{Gzip compression level. Default is 0 (no gzip compression).}
}
\value{
A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
\itemize{
\item \code{readBPCellsAnnHDF5Matrix}: read a sparce matrices from a AnnData HDF5
file on disk.
\item \code{writeBPCellsAnnHDF5Matrix}: Write a sparce matrices into a AnnData HDF5
file on disk.
}
}
\seealso{
\itemize{
\item \link{BPCellsSeed}
\item \link{BPCellsDir-IO}
\item \link{BPCellsHDF5-IO}
\item \link{BPCellsMem-IO}
\item \link{BPCells10xHDF5-IO}
\item \link{BPCellsAnnHDF5-IO}
}
}
