% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BindMatrix.R
\docType{class}
\name{BPCellsBindMatrix-Class}
\alias{BPCellsBindMatrix-Class}
\alias{BPCellsColBindMatrixArray-class}
\alias{DelayedArray,BPCellsColBindMatrixSeed-method}
\alias{BPCellsColBindMatrixArray}
\alias{BPCellsColBindMatrixMatrix-class}
\alias{matrixClass,BPCellsColBindMatrixArray-method}
\alias{BPCellsRowBindMatrixArray-class}
\alias{DelayedArray,BPCellsRowBindMatrixSeed-method}
\alias{BPCellsRowBindMatrixArray}
\alias{BPCellsRowBindMatrixMatrix-class}
\alias{matrixClass,BPCellsRowBindMatrixArray-method}
\title{Delayed BPCells ColBindMatrices and RowBindMatrices}
\usage{
\S4method{DelayedArray}{BPCellsColBindMatrixSeed}(seed)

BPCellsColBindMatrixArray(x)

\S4method{matrixClass}{BPCellsColBindMatrixArray}(x)

\S4method{DelayedArray}{BPCellsRowBindMatrixSeed}(seed)

BPCellsRowBindMatrixArray(x)

\S4method{matrixClass}{BPCellsRowBindMatrixArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsColBindMatrixSeed} or \code{BPCellsRowBindMatrixSeed}
object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsColBindMatrixArray} and \code{BPCellsRowBindMatrixArray}: A
\code{ColBindMatrices} or \code{RowBindMatrices} object.
\item \code{matrixClass}: A \code{BPCellsColBindMatrixArray} or \code{BPCellsRowBindMatrixArray}
object.
}}
}
\value{
\itemize{
\item \code{DelayedArray}: A \code{BPCellsColBindMatrixMatrix} or
\code{BPCellsRowBindMatrixMatrix} object.
}

\itemize{
\item \code{BPCellsColBindMatrixArray}: A \code{BPCellsColBindMatrixMatrix} object.
\item \code{BPCellsColBindMatrixArray}: A \code{BPCellsRowBindMatrixMatrix} object.
}
}
\description{
The \code{BPCellsRowBindMatrixArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{RowBindMatrices}
object in BPCells.

The \code{BPCellsColBindMatrixArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{ColBindMatrices}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[methods:cbind2]{cbind2} (\link[DelayedArray:DelayedArray-utils]{cbind},
\link[S4Arrays:abind]{acbind}, \link[S4Vectors:bindROWS]{bindCOLS}) or
\link[methods:cbind2]{rbind2} (\link[DelayedArray:DelayedArray-utils]{rbind},
\link[S4Arrays:abind]{arbind}, \link[S4Vectors:bindROWS]{bindROWS}) to create a
\code{BPCellsColBindMatrixMatrix} or \code{BPCellsRowBindMatrixMatrix} object.
}
\seealso{
\link{BPCellsSeed}
}
