% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BindMatrix.R
\docType{class}
\name{BPCellsBindMatrix}
\alias{BPCellsBindMatrix}
\alias{BPCellsColBindMatrixSeed}
\alias{BPCellsColBindMatrixArray-class}
\alias{DelayedArray,BPCellsColBindMatrixSeed-method}
\alias{BPCellsColBindMatrixArray}
\alias{BPCellsColBindMatrixMatrix-class}
\alias{matrixClass,BPCellsColBindMatrixArray-method}
\alias{[,BPCellsColBindMatrixSeed,ANY,ANY,ANY-method}
\alias{path,BPCellsColBindMatrixSeed-method}
\alias{BPCellsRowBindMatrixSeed-class}
\alias{BPCellsRowBindMatrixSeed}
\alias{BPCellsRowBindMatrixArray-class}
\alias{DelayedArray,BPCellsRowBindMatrixSeed-method}
\alias{BPCellsRowBindMatrixArray}
\alias{BPCellsRowBindMatrixMatrix-class}
\alias{matrixClass,BPCellsRowBindMatrixArray-method}
\alias{[,BPCellsRowBindMatrixSeed,ANY,ANY,ANY-method}
\alias{path,BPCellsRowBindMatrixSeed-method}
\title{Delayed BPCells ColBindMatrices and RowBindMatrices}
\usage{
BPCellsColBindMatrixSeed(x)

\S4method{DelayedArray}{BPCellsColBindMatrixSeed}(seed)

BPCellsColBindMatrixArray(x)

\S4method{matrixClass}{BPCellsColBindMatrixArray}(x)

\S4method{[}{BPCellsColBindMatrixSeed,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{path}{BPCellsColBindMatrixSeed}(object)

BPCellsRowBindMatrixSeed(x)

\S4method{DelayedArray}{BPCellsRowBindMatrixSeed}(seed)

BPCellsRowBindMatrixArray(x)

\S4method{matrixClass}{BPCellsRowBindMatrixArray}(x)

\S4method{[}{BPCellsRowBindMatrixSeed,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{path}{BPCellsRowBindMatrixSeed}(object)
}
\arguments{
\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsColBindMatrixSeed} and \code{BPCellsColBindMatrixArray}: A
\code{ColBindMatrices} object.
\item \code{BPCellsRowBindMatrixSeed} and \code{BPCellsRowBindMatrixArray}: A
\code{RowBindMatrices} object.
\item \code{[} and \code{path}: A \code{BPCellsColBindMatrixSeed} or \code{BPCellsRowBindMatrixSeed}
object.
}}

\item{seed}{A \code{BPCellsColBindMatrixSeed} or \code{BPCellsRowBindMatrixSeed}
object.}

\item{i, j}{Row and Column index.}

\item{...}{Ignored currently.}

\item{drop}{Ignored, always be \code{FALSE}.}

\item{object}{A \code{BPCellsColBindMatrixSeed} or \code{BPCellsRowBindMatrixSeed}
object.}
}
\value{
\itemize{
\item \code{BPCellsColBindMatrixSeed}: A \code{BPCellsColBindMatrixSeed} object.
\item \code{BPCellsRowBindMatrixSeed}: A \code{BPCellsRowBindMatrixSeed} object.
}

\itemize{
\item \code{DelayedArray}: A \code{BPCellsColBindMatrixMatrix} or
\code{BPCellsRowBindMatrixMatrix} object.
}

\itemize{
\item \code{BPCellsColBindMatrixArray}: A \code{BPCellsColBindMatrixMatrix} object.
\item \code{BPCellsColBindMatrixArray}: A \code{BPCellsRowBindMatrixMatrix} object.
}

\itemize{
\item \code{[}: Usually a \code{BPCellsSubsetSeed} object, but not always
}

\itemize{
\item \code{path}: A character file paths.
}
}
\description{
The \code{BPCellsRowBindMatrixArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{RowBindMatrices}
object in BPCells.

The \code{BPCellsColBindMatrixArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{ColBindMatrices}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[methods:cbind2]{cbind2} (\link[DelayedArray:DelayedArray-utils]{cbind},
\link[S4Arrays:abind]{acbind}, \link[S4Vectors:bindROWS]{bindCOLS}) or
\link[methods:cbind2]{rbind2} (\link[DelayedArray:DelayedArray-utils]{rbind},
\link[S4Arrays:abind]{arbind}, \link[S4Vectors:bindROWS]{bindROWS}) to create a
\code{BPCellsColBindMatrixMatrix} or \code{BPCellsRowBindMatrixMatrix} object.
}
