% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Convert.R
\docType{class}
\name{BPCellsConvert}
\alias{BPCellsConvert}
\alias{BPCellsConvertSeed}
\alias{BPCellsConvertArray-class}
\alias{DelayedArray,BPCellsConvertSeed-method}
\alias{BPCellsConvertArray}
\alias{BPCellsConvertMatrix-class}
\alias{matrixClass,BPCellsConvertArray-method}
\alias{path,BPCellsConvertSeed-method}
\alias{[,BPCellsConvertSeed,ANY,ANY,ANY-method}
\title{Delayed BPCells ConvertMatrixType}
\usage{
BPCellsConvertSeed(x)

\S4method{DelayedArray}{BPCellsConvertSeed}(seed)

BPCellsConvertArray(x)

\S4method{matrixClass}{BPCellsConvertArray}(x)

\S4method{path}{BPCellsConvertSeed}(object)

\S4method{[}{BPCellsConvertSeed,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{ConvertMatrixType} object.}

\item{seed}{A \code{BPCellsConvertSeed} object.}

\item{object}{A \code{BPCellsConvertSeed} object.}

\item{i, j}{Row and Column index.}

\item{...}{Ignored currently.}

\item{drop}{Ignored, always be \code{FALSE}.}
}
\description{
The \code{BPCellsConvertArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{ConvertMatrixType}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link{convert_type} to create a \code{BPCellsConvert} object.
}
