% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ConvertMatrixType.R
\docType{class}
\name{BPCellsConvertMatrixType}
\alias{BPCellsConvertMatrixType}
\alias{BPCellsConvertMatrixTypeSeed}
\alias{BPCellsConvertMatrixTypeArray-class}
\alias{DelayedArray,BPCellsConvertMatrixTypeSeed-method}
\alias{BPCellsConvertMatrixTypeArray}
\alias{BPCellsConvertMatrixTypeMatrix-class}
\alias{matrixClass,BPCellsConvertMatrixTypeArray-method}
\alias{show,BPCellsConvertMatrixTypeSeed-method}
\alias{type,BPCellsConvertMatrixTypeSeed-method}
\alias{is_sparse,BPCellsConvertMatrixTypeSeed-method}
\alias{path,BPCellsConvertMatrixTypeSeed-method}
\alias{extract_array,BPCellsConvertMatrixTypeSeed-method}
\alias{extract_sparse_array,BPCellsConvertMatrixTypeSeed-method}
\title{Delayed BPCells ConvertMatrixType}
\usage{
BPCellsConvertMatrixTypeSeed(x)

\S4method{DelayedArray}{BPCellsConvertMatrixTypeSeed}(seed)

BPCellsConvertMatrixTypeArray(x)

\S4method{matrixClass}{BPCellsConvertMatrixTypeArray}(x)

\S4method{show}{BPCellsConvertMatrixTypeSeed}(object)

\S4method{type}{BPCellsConvertMatrixTypeSeed}(x)

\S4method{is_sparse}{BPCellsConvertMatrixTypeSeed}(x)

\S4method{path}{BPCellsConvertMatrixTypeSeed}(object)

\S4method{extract_array}{BPCellsConvertMatrixTypeSeed}(x, index)

\S4method{extract_sparse_array}{BPCellsConvertMatrixTypeSeed}(x, index)
}
\arguments{
\item{x}{A \code{ConvertMatrixType} or \code{BPCellsConvertMatrixTypeSeed} object.}

\item{seed}{A \code{BPCellsConvertMatrixTypeSeed} object.}

\item{object}{A \code{BPCellsConvertMatrixTypeSeed} object.}

\item{index}{
    An unnamed list of integer vectors, one per dimension in \code{x}.
    Each vector is called a \emph{subscript} and can only contain
    positive integers that are valid 1-based indices along the corresponding
    dimension in \code{x}.

    Empty or missing subscripts are allowed. They must be represented
    by list elements set to \code{integer(0)} or \code{NULL}, respectively.

    The subscripts cannot contain NAs or non-positive values.

    Individual subscripts are allowed to contain duplicated indices.
  }
}
\description{
The \code{BPCellsConvertMatrixTypeArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{ConvertMatrixType}
object in BPCells. Usually, you shouldn't use this class directly, instead,
you should use \link{convert_type}.
}
