% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Dir.R
\docType{class}
\name{BPCellsDir}
\alias{BPCellsDir}
\alias{BPCellsDirArray-class}
\alias{DelayedArray,BPCellsDirSeed-method}
\alias{BPCellsDirArray}
\alias{BPCellsDirMatrix-class}
\alias{matrixClass,BPCellsDirArray-method}
\title{Delayed MatrixDir arrays}
\usage{
\S4method{DelayedArray}{BPCellsDirSeed}(seed)

BPCellsDirArray(x, ...)

\S4method{matrixClass}{BPCellsDirArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsDirSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsDirArray}: A \code{MatrixDir} object.
\item \code{matrixClass}: A \code{BPCellsDirArray} object.
}}

\item{...}{
  Arguments passed on to \code{\link[BPCells:matrix_io]{BPCells::open_matrix_dir}}
  \describe{
    \item{\code{buffer_size}}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}
  }}
}
\description{
The \code{BPCellsDirArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixDir} object in
BPCells.
}
\seealso{
\itemize{
\item \link{BPCellsSeed}
\item \link{writeBPCellsDirArray}
}
}
