% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO-HDF5.R
\name{BPCellsHDF5-IO}
\alias{BPCellsHDF5-IO}
\alias{readBPCellsHDF5Matrix}
\alias{writeBPCellsHDF5Matrix}
\alias{writeBPCellsHDF5Matrix,ANY-method}
\title{Read/write sparse matrices from (or into) HDF5 file}
\usage{
readBPCellsHDF5Matrix(path, group, buffer_size = 8192L, seedform = NULL)

writeBPCellsHDF5Matrix(x, ...)

\S4method{writeBPCellsHDF5Matrix}{ANY}(
  x,
  path,
  group,
  bitpacking = TRUE,
  buffer_size = 8192L,
  chunk_size = 1024L,
  overwrite = FALSE,
  gzip = 0L,
  seedform = NULL
)
}
\arguments{
\item{path}{A string path of the \code{HDF5} file to read or save data into.}

\item{group}{The group within the hdf5 file to write the data to. If writing
to an existing hdf5 file this group must not already be in use}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}

\item{seedform}{A string, \code{"BPCells"} or \code{"DelayedArray"}, if \code{NULL}, will
use the default value.
\itemize{
\item For \verb{readBPCells*}: (use \code{seedform()} to check) to check the default
value.
\item For \verb{writeBPCells*}:
\itemize{
\item For \code{BPCellsMatrix} object: the default value will be extracted from
\code{x} directly (use \code{seedform(x)} to check).
\item For other object: the default value will be extracted from global
option (use \code{seedform()} to check).
}
}}

\item{x}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object or any objects can be
converted into \link{BPCellsSeed} object.}

\item{...}{Additional arguments passed into specific methods.}

\item{bitpacking}{A bool, whether or not to compress the data using
Bitpacking Compression.}

\item{chunk_size}{For performance tuning only. The chunk size used for the HDF5 array storage.}

\item{overwrite}{A bool, If \code{TRUE}, write to a temp dir then overwrite
existing data.}

\item{gzip}{Gzip compression level. Default is 0 (no gzip compression). This
is recommended when both compression and compatibility with outside programs
is required. Using \code{compress=TRUE} is recommended as it is >10x faster with
often similar compression levels. So \code{gzip} will always be zero when
\code{compress} is \code{TRUE}.}
}
\value{
A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
\itemize{
\item \code{readBPCellsHDF5Matrix}: read a sparce matrices from a HDF5 file on disk
\item \code{writeBPCellsHDF5Matrix}: Write a sparce matrices into a HDF5 file on disk
}
}
\details{
\subsection{Storage locations}{

Matrices can be stored in a directory on disk, in memory, or in an HDF5 file.
Saving in a directory on disk is a good default for local analysis, as it provides
the best I/O performance and lowest memory usage. The HDF5 format
allows saving within existing hdf5 files to group data together, and the in
memory format provides the fastest performance in the event memory usage is
unimportant.
}

\subsection{Bitpacking Compression}{

For typical RNA counts matrices holding integer counts, this bitpacking
compression will result in 6-8x less space than an R dgCMatrix, and 4-6x
smaller than a scipy csc_matrix. The compression will be more effective when
the count values in the matrix are small, and when the rows of the matrix are
sorted by rowMeans. In tests on RNA-seq data optimal ordering could save up
to 40\% of storage space. On non-integer data only the row indices are
compressed, not the values themselves so space savings will be smaller.

For non-integer data matrices, bitpacking compression is much less effective,
as it can only be applied to the indexes of each entry but not the values.
There will still be some space savings, but far less than for counts matrices.
}
}
\seealso{
\itemize{
\item \link{BPCellsSeed}
\item \link{BPCellsDir-IO}
\item \link{BPCellsHDF5-IO}
\item \link{BPCellsMem-IO}
}
}
