% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix-Methods.R, R/Class-Transformed.R
\name{BPCellsMatrix-methods}
\alias{BPCellsMatrix-methods}
\alias{show,BPCellsMatrix-method}
\alias{t,BPCellsMatrix-method}
\alias{[,BPCellsMatrix,ANY,ANY,ANY-method}
\alias{[<-,BPCellsMatrix,ANY,ANY,ANY-method}
\alias{dimnames<-,BPCellsMatrix,ListOrNULL-method}
\alias{expm1_slow}
\alias{expm1_slow,BPCellsMatrix-method}
\alias{log1p_slow}
\alias{log1p_slow,BPCellsMatrix-method}
\alias{expm1,BPCellsMatrix-method}
\alias{expm1}
\alias{log1p,BPCellsMatrix-method}
\alias{log1p}
\alias{pmin_by_col}
\alias{pmin_by_col,BPCellsMatrix-method}
\alias{pmin_by_row}
\alias{pmin_by_row,BPCellsMatrix-method}
\alias{pmin_scalar}
\alias{pmin_scalar,BPCellsMatrix-method}
\alias{pow_slow}
\alias{pow_slow,BPCellsMatrix-method}
\alias{^,BPCellsMatrix,ANY-method}
\alias{^}
\alias{round,BPCellsMatrix-method}
\alias{round}
\title{Basic operations for \code{BPCellsMatrix} object}
\usage{
\S4method{show}{BPCellsMatrix}(object)

\S4method{t}{BPCellsMatrix}(x)

\S4method{[}{BPCellsMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{BPCellsMatrix,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{dimnames}{BPCellsMatrix,ListOrNULL}(x) <- value

expm1_slow(x)

\S4method{expm1_slow}{BPCellsMatrix}(x)

log1p_slow(x)

\S4method{log1p_slow}{BPCellsMatrix}(x)

\S4method{expm1}{BPCellsMatrix}(x)

\S4method{log1p}{BPCellsMatrix}(x)

pmin_by_col(object, values)

\S4method{pmin_by_col}{BPCellsMatrix}(object, values)

pmin_by_row(object, values)

\S4method{pmin_by_row}{BPCellsMatrix}(object, values)

pmin_scalar(object, value)

\S4method{pmin_scalar}{BPCellsMatrix}(object, value)

pow_slow(e1, e2)

\S4method{pow_slow}{BPCellsMatrix}(e1, e2)

\S4method{^}{BPCellsMatrix,ANY}(e1, e2)

\S4method{round}{BPCellsMatrix}(x, digits = 0)
}
\arguments{
\item{x, object}{A \code{BPCellsMatrix} object.}

\item{i, j}{Row and Column index.}

\item{...}{Not used currently.}

\item{drop}{Ignored, always be \code{FALSE}.}

\item{value}{\itemize{
\item \verb{dimnames<-}: A list of dimnames or \code{NULL}.
\item \verb{[<-}: A matrix which can be coerced into
\link[Matrix:dgCMatrix-class]{dgCMatrix}.
\item \code{pmin_scalar}: Single positive numeric value.
}}

\item{values}{A positive atomic numeric.}

\item{e1, e2}{One of \code{e1} or \code{e2} must be \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object, and the another must be a
number.}

\item{digits}{Integer indicating the number of decimal places}
}
\value{
\itemize{
\item \code{t}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \code{[}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \verb{[<-}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \verb{dimnames<-}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \code{expm1} and \code{expm1_slow}: compute \code{exp(x)-1} of matrix.
}

\itemize{
\item \code{log1p} and \code{log1p_slow}: compute \code{log(1+x)} of matrix.
}

\itemize{
\item \code{pmin_by_col}: Take the minimum with a per-col constant
}

\itemize{
\item \code{pmin_by_row}: Take the minimum with a per-row constant
}

\itemize{
\item \code{pmin_scalar}: Take minumum with a global constant
}

\itemize{
\item \code{round}: Rounding of matrix Numbers.
}
}
\description{
Basic operations for \code{BPCellsMatrix} object
}
\seealso{
\itemize{
\item \link[=BPCells-bind]{bind}: Combine two Objects by Columns or Rows.
\item \link[=BPCells-Multiplication]{\%*\%}: Matrix Multiplication.
\item \link[=BPCells-crossprod]{crossprod}: Matrix Crossproduct.
\item \link[=BPCells-Summarization]{summarization}: row/col summarization.
\item \link[=BPCells-Arithmetic]{Arithmetic}: Binary Arithmetic operators.
\item \link[=BPCells-binarize]{binarize}: Convert matrix elements to zeros and ones.
}
}
