% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BPCellsMatrix.R
\docType{class}
\name{BPCellsMatrix}
\alias{BPCellsMatrix}
\alias{show,BPCellsMatrix-method}
\alias{t,BPCellsMatrix-method}
\title{Base Class for Delayed BPCells matrix}
\usage{
\S4method{show}{BPCellsMatrix}(object)

\S4method{t}{BPCellsMatrix}(x)

\S4method{\%*\%}{BPCellsMatrix,BPCellsMatrix}(x, y)
}
\arguments{
\item{object}{A \code{BPCellsMatrix} object.}
}
\value{
\itemize{
\item \code{x \%*\% y}: Matrix products, a \link{BPCellsMatrixMultiplyMatrix} object
}
}
\description{
The \code{BPCellsMatrix} class just inherits from the
\link[DelayedArray:DelayedArray-class]{DelayedMatrix} object. The purpose for
\code{BPCellsMatrix} object is to provide the common methods for all Delayed
BPCells matrix.
}
