% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MatrixSubset.R
\docType{class}
\name{BPCellsMatrixSubset}
\alias{BPCellsMatrixSubset}
\alias{BPCellsMatrixSubsetSeed}
\alias{BPCellsMatrixSubsetArray-class}
\alias{DelayedArray,BPCellsMatrixSubsetSeed-method}
\alias{BPCellsMatrixSubsetArray}
\alias{BPCellsMatrixSubsetMatrix-class}
\alias{matrixClass,BPCellsMatrixSubsetArray-method}
\alias{show,BPCellsMatrixSubsetSeed-method}
\alias{type,BPCellsMatrixSubsetSeed-method}
\alias{is_sparse,BPCellsMatrixSubsetSeed-method}
\alias{path,BPCellsMatrixSubsetSeed-method}
\alias{extract_array,BPCellsMatrixSubsetSeed-method}
\alias{extract_sparse_array,BPCellsMatrixSubsetSeed-method}
\title{Delayed BPCells MatrixSubset}
\usage{
BPCellsMatrixSubsetSeed(x)

\S4method{DelayedArray}{BPCellsMatrixSubsetSeed}(seed)

BPCellsMatrixSubsetArray(x)

\S4method{matrixClass}{BPCellsMatrixSubsetArray}(x)

\S4method{show}{BPCellsMatrixSubsetSeed}(object)

\S4method{type}{BPCellsMatrixSubsetSeed}(x)

\S4method{is_sparse}{BPCellsMatrixSubsetSeed}(x)

\S4method{path}{BPCellsMatrixSubsetSeed}(object)

\S4method{extract_array}{BPCellsMatrixSubsetSeed}(x, index)

\S4method{extract_sparse_array}{BPCellsMatrixSubsetSeed}(x, index)
}
\arguments{
\item{x}{A \code{MatrixSubset} or \code{BPCellsMatrixSubsetSeed} object.}

\item{seed}{A \code{BPCellsMatrixSubsetSeed} object.}

\item{object}{A \code{BPCellsMatrixSubsetSeed} object.}

\item{index}{
    An unnamed list of integer vectors, one per dimension in \code{x}.
    Each vector is called a \emph{subscript} and can only contain
    positive integers that are valid 1-based indices along the corresponding
    dimension in \code{x}.

    Empty or missing subscripts are allowed. They must be represented
    by list elements set to \code{integer(0)} or \code{NULL}, respectively.

    The subscripts cannot contain NAs or non-positive values.

    Individual subscripts are allowed to contain duplicated indices.
  }
}
\description{
The \code{BPCellsMatrixSubsetArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixSubset} object
in BPCells. Usually, you shouldn't use this class directly, instead, you
should use \code{[} (extract methods) of other BPCellsArray objects.
}
