% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO-Mem.R
\name{BPCellsMem-IO}
\alias{BPCellsMem-IO}
\alias{writeBPCellsMemMatrix}
\alias{writeBPCellsMemMatrix,ANY-method}
\title{Write a sparce matrices into memory with BPCells format}
\usage{
writeBPCellsMemMatrix(x, ...)

\S4method{writeBPCellsMemMatrix}{ANY}(x, compress = TRUE)
}
\arguments{
\item{x}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object or a matrix-like object which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix} object.}

\item{...}{Additional arguments passed into specific methods.}

\item{compress}{Whether or not to compress the data.}
}
\value{
A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
Write a sparce matrices into memory with BPCells format
}
\details{
\subsection{Storage locations}{

Matrices can be stored in a directory on disk, in memory, or in an HDF5 file.
Saving in a directory on disk is a good default for local analysis, as it provides
the best I/O performance and lowest memory usage. The HDF5 format
allows saving within existing hdf5 files to group data together, and the in
memory format provides the fastest performance in the event memory usage is
unimportant.
}

\subsection{Bitpacking Compression}{

For typical RNA counts matrices holding integer counts, this bitpacking
compression will result in 6-8x less space than an R dgCMatrix, and 4-6x
smaller than a scipy csc_matrix. The compression will be more effective when
the count values in the matrix are small, and when the rows of the matrix are
sorted by rowMeans. In tests on RNA-seq data optimal ordering could save up
to 40\% of storage space. On non-integer data only the row indices are
compressed, not the values themselves so space savings will be smaller.

For non-integer data matrices, bitpacking compression is much less effective,
as it can only be applied to the indexes of each entry but not the values.
There will still be some space savings, but far less than for counts matrices.
}
}
\seealso{
\itemize{
\item \link{BPCellsSeed}
\item \link{BPCellsDir-IO}
\item \link{BPCellsHDF5-IO}
\item \link{BPCellsMem-IO}
\item \link{BPCells10xHDF5-IO}
\item \link{BPCellsAnnHDF5-IO}
}
}
