% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Mem.R
\docType{class}
\name{BPCellsMem}
\alias{BPCellsMem}
\alias{BPCellsMemArray-class}
\alias{DelayedArray,BPCellsMemSeed-method}
\alias{BPCellsMemArray}
\alias{BPCellsMemMatrix-class}
\alias{matrixClass,BPCellsMemArray-method}
\title{Delayed PackedMatrixMemBase and UnpackedMatrixMemBase of BPCells}
\usage{
\S4method{DelayedArray}{BPCellsMemSeed}(seed)

BPCellsMemArray(x)

\S4method{matrixClass}{BPCellsMemArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsMemSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsMemArray}: A \code{PackedMatrixMemBase} or \code{UnpackedMatrixMemBase}
object.
\item \code{matrixClass}: A \code{BPCellsMemArray} object.
}}
}
\description{
The \code{BPCellsMemArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{PackedMatrixMemBase}
and \code{UnpackedMatrixMemBase} object.
}
\seealso{
\itemize{
\item \link{BPCellsSeed}
\item \link{writeBPCellsMemArray}
}
}
