% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Multiply.R
\docType{class}
\name{BPCellsMultiply}
\alias{BPCellsMultiply}
\alias{BPCellsMultiplySeed}
\alias{BPCellsMultiplyArray-class}
\alias{DelayedArray,BPCellsMultiplySeed-method}
\alias{BPCellsMultiplyArray}
\alias{BPCellsMultiplyMatrix-class}
\alias{matrixClass,BPCellsMultiplyArray-method}
\alias{path,BPCellsMultiplySeed-method}
\title{Delayed BPCells MatrixMultiply}
\usage{
BPCellsMultiplySeed(x)

\S4method{DelayedArray}{BPCellsMultiplySeed}(seed)

BPCellsMultiplyArray(x)

\S4method{matrixClass}{BPCellsMultiplyArray}(x)

\S4method{path}{BPCellsMultiplySeed}(object)
}
\arguments{
\item{x}{A \code{MatrixMultiply} object.}

\item{seed}{A \code{BPCellsMultiplySeed} object.}
}
\description{
The \code{BPCellsMultiplyArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixMultiply}
object in BPCells.
}
\details{
Usually, you shouldn't use this class directly, instead, you should use \code{\%*\%}
or \code{crossprod} methods of other BPCellsMatrix objects.
}
