% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-RankTransform.R
\docType{class}
\name{BPCellsRankTransform}
\alias{BPCellsRankTransform}
\alias{BPCellsRankTransformArray-class}
\alias{DelayedArray,BPCellsRankTransformSeed-method}
\alias{BPCellsRankTransformArray}
\alias{BPCellsRankTransformMatrix-class}
\alias{matrixClass,BPCellsRankTransformArray-method}
\title{Delayed BPCells MatrixRankTransform}
\usage{
\S4method{DelayedArray}{BPCellsRankTransformSeed}(seed)

BPCellsRankTransformArray(x)

\S4method{matrixClass}{BPCellsRankTransformArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsRankTransformSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsMatrixRankTransformArray}: A \code{MatrixRankTransform} object.
\item \code{matrixClass}: A \code{BPCellsMatrixRankTransformArray} object.
}}
}
\description{
The \code{BPCellsRankTransformArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixRankTransform}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link{rank_transform} function to create a \code{BPCellsRankTransformMatrix}.
}
\seealso{
\itemize{
\item \link{BPCellsSeed}
\item \link[BPCells:convert_matrix_type]{convert_matrix_type}
}
}
