% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-RenameDims.R
\docType{class}
\name{BPCellsRenameDims}
\alias{BPCellsRenameDims}
\alias{BPCellsRenameDimsSeed}
\alias{BPCellsRenameDimsArray-class}
\alias{DelayedArray,BPCellsRenameDimsSeed-method}
\alias{BPCellsRenameDimsArray}
\alias{BPCellsRenameDimsMatrix-class}
\alias{matrixClass,BPCellsRenameDimsArray-method}
\alias{path,BPCellsRenameDimsSeed-method}
\alias{[,BPCellsRenameDimsSeed,ANY,ANY,ANY-method}
\title{Delayed BPCells RenameDims}
\usage{
BPCellsRenameDimsSeed(x)

\S4method{DelayedArray}{BPCellsRenameDimsSeed}(seed)

BPCellsRenameDimsArray(x)

\S4method{matrixClass}{BPCellsRenameDimsArray}(x)

\S4method{path}{BPCellsRenameDimsSeed}(object)

\S4method{[}{BPCellsRenameDimsSeed,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{RenameDims} object.}

\item{seed}{A \code{BPCellsRenameDimsSeed} object.}

\item{object}{A \code{BPCellsRenameDimsSeed} object.}

\item{i, j}{Row and Column index.}

\item{drop}{Not used, always be \code{FALSE}.}
}
\description{
The \code{BPCellsRenameDimsArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{RenameDims}
object in BPCells.
}
\details{
Usually, you shouldn't use this class directly, instead, you should use
\link{dimnames<-} to create a \link[=BPCellsRenameDims-Class]{BPCellsRenameDims} object.
}
