% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPCellsSeed.R, R/Class-Convert.R
\name{BPCellsSeed}
\alias{BPCellsSeed}
\alias{BPCellsSeed,IterableMatrix-method}
\alias{BPCellsSeed,matrix-method}
\alias{BPCellsSeed,dgCMatrix-method}
\alias{BPCellsSeed,ANY-method}
\alias{summary.BPCellsDelayedConvert}
\title{Build \code{BPCellsSeed} object}
\usage{
BPCellsSeed(x)

\S4method{BPCellsSeed}{IterableMatrix}(x)

\S4method{BPCellsSeed}{matrix}(x)

\S4method{BPCellsSeed}{dgCMatrix}(x)

\S4method{BPCellsSeed}{ANY}(x)

summary.BPCellsDelayedConvert(object)
}
\arguments{
\item{x}{A \code{IterableMatrix} object from \code{BPCells}, a matrix-like object which
can be coerced into dgCMatrix object.}
}
\value{
A \link[=BPCellsSeed-class]{BPCellsSeed} object.
}
\description{
Build \code{BPCellsSeed} object
}
\seealso{
\linkS4class{BPCellsSeed}
}
