% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-Binarize.R
\docType{class}
\name{BPCellsTransformBinarize}
\alias{BPCellsTransformBinarize}
\alias{BPCellsTransformBinarizeArray-class}
\alias{DelayedArray,BPCellsTransformBinarizeSeed-method}
\alias{BPCellsTransformBinarizeArray}
\alias{BPCellsTransformBinarizeMatrix-class}
\alias{matrixClass,BPCellsTransformBinarizeArray-method}
\title{Delayed BPCells TransformBinarize}
\usage{
\S4method{DelayedArray}{BPCellsTransformBinarizeSeed}(seed)

BPCellsTransformBinarizeArray(x)

\S4method{matrixClass}{BPCellsTransformBinarizeArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformBinarizeSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformBinarizeArray}: A \code{TransformBinarize} object.
\item \code{matrixClass}: A \code{BPCellsTransformBinarizeArray} object.
}}
}
\value{
A \code{BPCellsTransformBinarizeMatrix} object.
}
\description{
The \code{BPCellsTransformBinarizeArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformBinarize}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCells-binarize]{binarize} to create \code{BPCellsTransformBinarizeMatrix}
object.
}
\seealso{
\link{BPCellsSeed}
}
