% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Transformed.R
\docType{class}
\name{BPCellsTransformed}
\alias{BPCellsTransformed}
\alias{BPCellsTransformedArray-class}
\alias{DelayedArray,BPCellsTransformedSeed-method}
\alias{BPCellsTransformedArray}
\alias{BPCellsTransformedMatrix-class}
\alias{matrixClass,BPCellsTransformedArray-method}
\title{Delayed BPCells TransformedMatrix}
\usage{
\S4method{DelayedArray}{BPCellsTransformedSeed}(seed)

BPCellsTransformedArray(x)

\S4method{matrixClass}{BPCellsTransformedArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformedSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformedMatrixArray}: A \code{TransformedMatrix} object.
\item \code{matrixClass}: A \code{BPCellsTransformedMatrixArray} object.
}}
}
\description{
The \code{BPCellsTransformedArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformedMatrix}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\code{[} (extract methods) of other BPCellsArray objects.
}
\seealso{
\link{BPCellsSeed}
}
