% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-expm1.R
\docType{class}
\name{BPCellsTransformExpm1}
\alias{BPCellsTransformExpm1}
\alias{BPCellsTransformExpm1Array-class}
\alias{DelayedArray,BPCellsTransformExpm1Seed-method}
\alias{BPCellsTransformExpm1Array}
\alias{BPCellsTransformExpm1Matrix-class}
\alias{matrixClass,BPCellsTransformExpm1Array-method}
\title{Delayed BPCells TransformExpm1}
\usage{
\S4method{DelayedArray}{BPCellsTransformExpm1Seed}(seed)

BPCellsTransformExpm1Array(x)

\S4method{matrixClass}{BPCellsTransformExpm1Array}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformExpm1Seed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformExpm1Array}: A \code{TransformExpm1} object.
\item \code{matrixClass}: A \code{BPCellsTransformExpm1Array} object.
}}
}
\value{
A \code{BPCellsTransformExpm1Matrix} object.
}
\description{
The \code{BPCellsTransformExpm1Array} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformExpm1}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{expm1} to create \code{BPCellsTransformExpm1Matrix}
object.
}
