% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-expm1_slow.R
\docType{class}
\name{BPCellsTransformExpm1Slow}
\alias{BPCellsTransformExpm1Slow}
\alias{BPCellsTransformExpm1SlowArray-class}
\alias{DelayedArray,BPCellsTransformExpm1SlowSeed-method}
\alias{BPCellsTransformExpm1SlowArray}
\alias{BPCellsTransformExpm1SlowMatrix-class}
\alias{matrixClass,BPCellsTransformExpm1SlowArray-method}
\title{Delayed BPCells TransformExpm1Slow}
\usage{
\S4method{DelayedArray}{BPCellsTransformExpm1SlowSeed}(seed)

BPCellsTransformExpm1SlowArray(x)

\S4method{matrixClass}{BPCellsTransformExpm1SlowArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformExpm1SlowSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformExpm1SlowArray}: A \code{TransformExpm1Slow} object.
\item \code{matrixClass}: A \code{BPCellsTransformExpm1SlowArray} object.
}}
}
\description{
The \code{BPCellsTransformExpm1SlowArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformExpm1Slow}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{expm1_slow} to create \code{BPCellsTransformExpm1SlowMatrix}
object.
}
