% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-log1p.R
\docType{class}
\name{BPCellsTransformLog1p}
\alias{BPCellsTransformLog1p}
\alias{BPCellsTransformLog1pArray-class}
\alias{DelayedArray,BPCellsTransformLog1pSeed-method}
\alias{BPCellsTransformLog1pArray}
\alias{BPCellsTransformLog1pMatrix-class}
\alias{matrixClass,BPCellsTransformLog1pArray-method}
\title{Delayed BPCells TransformLog1p}
\usage{
\S4method{DelayedArray}{BPCellsTransformLog1pSeed}(seed)

BPCellsTransformLog1pArray(x)

\S4method{matrixClass}{BPCellsTransformLog1pArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformLog1pSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformLog1pArray}: A \code{TransformLog1p} object.
\item \code{matrixClass}: A \code{BPCellsTransformLog1pArray} object.
}}
}
\value{
A \code{BPCellsTransformLog1pMatrix} object.
}
\description{
The \code{BPCellsTransformLog1pArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformLog1p}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{log1p} to create \code{BPCellsTransformLog1pMatrix}
object.
}
