% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-log1p_slow.R
\docType{class}
\name{BPCellsTransformLog1pSlow}
\alias{BPCellsTransformLog1pSlow}
\alias{BPCellsTransformLog1pSlowArray-class}
\alias{DelayedArray,BPCellsTransformLog1pSlowSeed-method}
\alias{BPCellsTransformLog1pSlowArray}
\alias{BPCellsTransformLog1pSlowMatrix-class}
\alias{matrixClass,BPCellsTransformLog1pSlowArray-method}
\title{Delayed BPCells TransformLog1pSlow}
\usage{
\S4method{DelayedArray}{BPCellsTransformLog1pSlowSeed}(seed)

BPCellsTransformLog1pSlowArray(x)

\S4method{matrixClass}{BPCellsTransformLog1pSlowArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformLog1pSlowSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformLog1pSlowArray}: A \code{TransformLog1pSlow} object.
\item \code{matrixClass}: A \code{BPCellsTransformLog1pSlowArray} object.
}}
}
\value{
A \code{BPCellsTransformLog1pSlowMatrix} object.
}
\description{
The \code{BPCellsTransformLog1pSlowArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformLog1pSlow}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{log1p_slow} to create \code{BPCellsTransformLog1pSlowMatrix}
object.
}
