% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-min_scalar.R
\docType{class}
\name{BPCellsTransformMin}
\alias{BPCellsTransformMin}
\alias{BPCellsTransformMinArray-class}
\alias{DelayedArray,BPCellsTransformMinSeed-method}
\alias{BPCellsTransformMinArray}
\alias{BPCellsTransformMinMatrix-class}
\alias{matrixClass,BPCellsTransformMinArray-method}
\title{Delayed BPCells TransformMin}
\usage{
\S4method{DelayedArray}{BPCellsTransformMinSeed}(seed)

BPCellsTransformMinArray(x)

\S4method{matrixClass}{BPCellsTransformMinArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformMinSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformMinArray}: A \code{TransformMin} object.
\item \code{matrixClass}: A \code{BPCellsTransformMinArray} object.
}}
}
\value{
A \code{BPCellsTransformMinMatrix} object.
}
\description{
The \code{BPCellsTransformMinArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformMin}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{pmin_scalar} to create \code{BPCellsTransformMinMatrix}
object.
}
