% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-min_by_col.R
\docType{class}
\name{BPCellsTransformMinByCol}
\alias{BPCellsTransformMinByCol}
\alias{BPCellsTransformMinByColArray-class}
\alias{DelayedArray,BPCellsTransformMinByColSeed-method}
\alias{BPCellsTransformMinByColArray}
\alias{BPCellsTransformMinByColMatrix-class}
\alias{matrixClass,BPCellsTransformMinByColArray-method}
\title{Delayed BPCells TransformMinByCol}
\usage{
\S4method{DelayedArray}{BPCellsTransformMinByColSeed}(seed)

BPCellsTransformMinByColArray(x)

\S4method{matrixClass}{BPCellsTransformMinByColArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformMinByColSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformMinByColArray}: A \code{TransformMinByCol} object.
\item \code{matrixClass}: A \code{BPCellsTransformMinByColArray} object.
}}
}
\value{
A \code{BPCellsTransformMinByColMatrix} object.
}
\description{
The \code{BPCellsTransformMinByColArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformMinByCol}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{pmin_by_col} to create \code{BPCellsTransformMinByColMatrix}
object.
}
