% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-min_by_row.R
\docType{class}
\name{BPCellsTransformMinByRow}
\alias{BPCellsTransformMinByRow}
\alias{BPCellsTransformMinByRowArray-class}
\alias{DelayedArray,BPCellsTransformMinByRowSeed-method}
\alias{BPCellsTransformMinByRowArray}
\alias{BPCellsTransformMinByRowMatrix-class}
\title{Delayed BPCells TransformMinByRow}
\usage{
\S4method{DelayedArray}{BPCellsTransformMinByRowSeed}(seed)

BPCellsTransformMinByRowArray(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformMinByRowSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformMinByRowArray}: A \code{TransformMinByRow} object.
\item \code{matrixClass}: A \code{BPCellsTransformMinByRowArray} object.
}}
}
\value{
A \code{BPCellsTransformMinByRowMatrix} object.
}
\description{
The \code{BPCellsTransformMinByRowArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformMinByRow}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{pmin_by_row} to create \code{BPCellsTransformMinByRowMatrix}
object.
}
