% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-pow.R
\docType{class}
\name{BPCellsTransformPow}
\alias{BPCellsTransformPow}
\alias{BPCellsTransformPowArray-class}
\alias{DelayedArray,BPCellsTransformPowSeed-method}
\alias{BPCellsTransformPowArray}
\alias{BPCellsTransformPowMatrix-class}
\alias{matrixClass,BPCellsTransformPowArray-method}
\title{Delayed BPCells TransformPow}
\usage{
\S4method{DelayedArray}{BPCellsTransformPowSeed}(seed)

BPCellsTransformPowArray(x)

\S4method{matrixClass}{BPCellsTransformPowArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformPowSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformPowArray}: A \code{TransformPow} object.
\item \code{matrixClass}: A \code{BPCellsTransformPowArray} object.
}}
}
\value{
A \code{BPCellsTransformPowMatrix} object.
}
\description{
The \code{BPCellsTransformPowArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformPow}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{^} to create \code{BPCellsTransformPowMatrix} object.
}
