% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-pow_slow.R
\docType{class}
\name{BPCellsTransformPowSlow}
\alias{BPCellsTransformPowSlow}
\alias{BPCellsTransformPowSlowArray-class}
\alias{DelayedArray,BPCellsTransformPowSlowSeed-method}
\alias{BPCellsTransformPowSlowArray}
\alias{BPCellsTransformPowSlowMatrix-class}
\alias{matrixClass,BPCellsTransformPowSlowArray-method}
\title{Delayed BPCells TransformPowSlow}
\usage{
\S4method{DelayedArray}{BPCellsTransformPowSlowSeed}(seed)

BPCellsTransformPowSlowArray(x)

\S4method{matrixClass}{BPCellsTransformPowSlowArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformPowSlowSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformPowSlowArray}: A \code{TransformPowSlow} object.
\item \code{matrixClass}: A \code{BPCellsTransformPowSlowArray} object.
}}
}
\value{
A \code{BPCellsTransformPowSlowMatrix} object.
}
\description{
The \code{BPCellsTransformPowSlowArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformPowSlow}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{pow_slow} to create \code{BPCellsTransformPowSlowMatrix}
object.
}
