% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-Round.R
\docType{class}
\name{BPCellsTransformRound}
\alias{BPCellsTransformRound}
\alias{BPCellsTransformRoundArray-class}
\alias{DelayedArray,BPCellsTransformRoundSeed-method}
\alias{BPCellsTransformRoundArray}
\alias{BPCellsTransformRoundMatrix-class}
\alias{matrixClass,BPCellsTransformRoundArray-method}
\title{Delayed BPCells TransformRound}
\usage{
\S4method{DelayedArray}{BPCellsTransformRoundSeed}(seed)

BPCellsTransformRoundArray(x)

\S4method{matrixClass}{BPCellsTransformRoundArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformRoundSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformRoundArray}: A \code{TransformRound} object.
\item \code{matrixClass}: A \code{BPCellsTransformRoundArray} object.
}}
}
\value{
A \code{BPCellsTransformRoundMatrix} object.
}
\description{
The \code{BPCellsTransformRoundArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformRound}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{round} to create \code{BPCellsTransformRoundMatrix} object.
}
