% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-Arithmetic.R
\docType{class}
\name{BPCellsTransformScaleShift}
\alias{BPCellsTransformScaleShift}
\alias{BPCellsTransformScaleShiftArray-class}
\alias{DelayedArray,BPCellsTransformScaleShiftSeed-method}
\alias{BPCellsTransformScaleShiftArray}
\alias{BPCellsTransformScaleShiftMatrix-class}
\alias{matrixClass,BPCellsTransformScaleShiftArray-method}
\title{Delayed BPCells TransformScaleShift}
\usage{
\S4method{DelayedArray}{BPCellsTransformScaleShiftSeed}(seed)

BPCellsTransformScaleShiftArray(x)

\S4method{matrixClass}{BPCellsTransformScaleShiftArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformScaleShiftSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformScaleShiftArray}: A \code{TransformScaleShift} object.
\item \code{matrixClass}: A \code{BPCellsTransformScaleShiftArray} object.
}}
}
\value{
A \code{BPCellsTransformScaleShiftMatrix} object.
}
\description{
The \code{BPCellsTransformScaleShiftArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformScaleShift}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\code{[} (extract methods) of other BPCellsArray objects.

Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{round} to create \code{BPCellsTransformScaleShiftMatrix} object.
}
