% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform-square.R
\docType{class}
\name{BPCellsTransformSquare}
\alias{BPCellsTransformSquare}
\alias{BPCellsTransformSquareArray-class}
\alias{DelayedArray,BPCellsTransformSquareSeed-method}
\alias{BPCellsTransformSquareArray}
\alias{BPCellsTransformSquareMatrix-class}
\alias{matrixClass,BPCellsTransformSquareArray-method}
\title{Delayed BPCells TransformSquare}
\usage{
\S4method{DelayedArray}{BPCellsTransformSquareSeed}(seed)

BPCellsTransformSquareArray(x)

\S4method{matrixClass}{BPCellsTransformSquareArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsTransformSquareSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsTransformSquareArray}: A \code{TransformSquare} object.
\item \code{matrixClass}: A \code{BPCellsTransformSquareArray} object.
}}
}
\value{
A \code{BPCellsTransformSquareMatrix} object.
}
\description{
The \code{BPCellsTransformSquareArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{TransformSquare}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\code{[} (extract methods) of other BPCellsArray objects.

Usually, you shouldn't use this class directly, instead, you should use
\link[=BPCellsMatrix-class]{round} to create \code{BPCellsTransformSquareMatrix} object.
}
