% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Mask.R
\name{mask_matrix}
\alias{mask_matrix}
\alias{mask_matrix,BPCellsMatrix-method}
\title{Mask matrix entries to zero}
\usage{
mask_matrix(object, mask, ...)

\S4method{mask_matrix}{BPCellsMatrix}(object, mask, invert = FALSE)
}
\arguments{
\item{object}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{mask}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object or any objects can
be converted into \link{BPCellsSeed} object.}

\item{...}{Additional parameters passed into specific methods.}

\item{invert}{A bool, indicates whether revert the mask.}
}
\value{
A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
Set matrix entries to zero given a mask matrix of the same dimensions.
Normally, non-zero values in the mask will set the matrix entry to zero. If
inverted, zero values in the mask matrix will set the matrix entry to zero.
}
\seealso{
\link[BPCells:mask_matrix]{mask_matrix}
}
