% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Dir.R
\name{readDirMatrix}
\alias{readDirMatrix}
\title{Read a BPCells Dir matrix from the path}
\usage{
readDirMatrix(path, ...)
}
\arguments{
\item{path}{A string for the path of the BPCells matrix dir.}

\item{...}{
  Arguments passed on to \code{\link[BPCells:matrix_io]{BPCells::open_matrix_dir}}
  \describe{
    \item{\code{buffer_size}}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}
  }}
}
\description{
Read a BPCells Dir matrix from the path
}
