% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seed-management.R, R/Class-BPCellsMatrix.R
\name{seedform}
\alias{seedform}
\alias{seedform,missing-method}
\alias{seedform,character-method}
\alias{seedform,BPCellsMatrix-method}
\alias{seedform<-}
\alias{seedform<-,BPCellsMatrix-method}
\title{BPCellsArray seed management}
\usage{
seedform(x)

\S4method{seedform}{missing}(x)

\S4method{seedform}{character}(x)

\S4method{seedform}{BPCellsMatrix}(x)

seedform(x, ...) <- value

\S4method{seedform}{BPCellsMatrix}(x) <- value
}
\arguments{
\item{x}{See \code{Value} section}

\item{...}{Additional argumentds passed into specific methods.}

\item{value}{A string, \code{"BPCells"} or \code{"DelayedArray"}.}
}
\value{
\itemize{
\item \code{missing}: Get current global \code{seedform} value.
\item \code{character}: Change the global \code{seedform} value and return the original
global \code{seedform} value invisiblely.
\item \code{BPCellsMatrix}:
\itemize{
\item For \code{seedform} method: Get the seedform of \link{BPCellsMatrix} object.
\item For \verb{seedform<-} method: Change the \code{seedform} value of a
\link{BPCellsMatrix} object.
}
}
}
\description{
\code{BPCellsArray} offers two seed form to use BPCells Matrix as the DelayedArray
seed:
\itemize{
\item \code{BPCells}: use the \code{IterableMatrix} of BPCells as the seed directly.
\item \code{DelayedArray}: convert \code{IterableMatrix} into a parallel
\link[DelayedArray:DelayedOp-class]{DelayedOp} object
(See \link[=BPCellsDelayedOp-class]{BPCellsDelayedOp} object).
}

Both methods are generally effective for most operations. However, choosing
\code{seedform=DelayedArray} ensures better compatibility with DelayedArray
package. It should be noted that \link[DelayedArray:showtree]{showtree} and
\link[DelayedArray:showtree]{seedApply} from DelayedArray do not function well
when \code{seedform=BPCells}. Nonetheless, BPCellsArray provides its own
\link{seedApply} and \link{showtree} functions for both methods. Overall, executing
operations with \code{seedform=BPCells} is expected to be faster due to the
absence of the need to convert between \code{IterableMatrix} and
\link[=BPCellsDelayedOp-class]{BPCellsDelayedOp} object. \code{seedform=BPCells} is also
the default method.
}
\section{Default seedform}{
 All function in \code{BPCellsArray} will use the global
\code{seedform} value as the default seedform (use \code{seedform()} to check), except
for \code{BPCellsMatrix} object, in which the default will be extracted from the
object directly (use \code{seedform(object)} to check).
}

