% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delayed-management.R
\name{set_delayed}
\alias{set_delayed}
\alias{set_delayed,missing-method}
\alias{set_delayed,logical-method}
\alias{set_delayed,BPCellsMatrix-method}
\title{BPCellsArray seed management}
\usage{
set_delayed(x, ...)

\S4method{set_delayed}{missing}(x)

\S4method{set_delayed}{logical}(x)

\S4method{set_delayed}{BPCellsMatrix}(x, delayed = NULL)
}
\arguments{
\item{x}{See return section}

\item{...}{Additional argumentds passed into specific methods.}

\item{delayed}{A bool value.}
}
\value{
\itemize{
\item \code{missing}: Get current default \code{delayed} value.
\item \code{logical}: Change the default \code{delayed} value and return original default
\code{delayed} value invisiblely.
\item \code{BPCellsMatrix}: Change the \code{delayed} value of a \link{BPCellsMatrix} object.
}
}
\description{
\code{BPCellsArray} offers two methods for using BPCells Matrix as the
DelayedArray seed:
\itemize{
\item \code{delayed=FALSE}: use the \code{IterableMatrix} of BPCells as the seed directly.
\item \code{delayed=TRUE}: convert \code{IterableMatrix} into a parallel
\link[DelayedArray:DelayedOp-class]{DelayedOp} object (\code{BPCellsDelayedOp}).
}

Both methods are generally effective for most operations. However, choosing
\code{delayed=TRUE} ensures better compatibility with \code{DelayedArray}. It should be
noted that \link[DelayedArray:showtree]{showtree} and
\link[DelayedArray:showtree]{seedApply} from DelayedArray do not function well
when \code{delayed=FALSE}. Nonetheless, BPCellsArray provides its own \link{seedApply}
and \link{showtree} functions for both methods. Overall, executing operations with
\code{delayed=FALSE} is expected to be faster due to the absence of the need to
convert between \code{IterableMatrix} and \code{BPCellsDelayedOp} objects.
\code{delayed=FALSE} is also the default method.
}
