% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seed-management.R, R/Class-BPCellsMatrix.R
\name{set_seed_form}
\alias{set_seed_form}
\alias{set_seed_form,missing-method}
\alias{set_seed_form,character-method}
\alias{set_seed_form,BPCellsMatrix-method}
\title{BPCellsArray seed management}
\usage{
set_seed_form(x, ...)

\S4method{set_seed_form}{missing}(x)

\S4method{set_seed_form}{character}(x)

\S4method{set_seed_form}{BPCellsMatrix}(x, seed_form = NULL)
}
\arguments{
\item{x}{See return section}

\item{...}{Additional argumentds passed into specific methods.}

\item{seed_form}{A string, "BPCells" or "DelayedArray", if \code{NULL}, will use
the default value (use \code{set_seed_form()} with missing argument to check).}
}
\value{
\itemize{
\item \code{missing}: Get current default \code{seed_form} value.
\item \code{character}: Change the default \code{seed_form} value and return the original
default \code{seed_form} value invisiblely.
\item \code{BPCellsMatrix}: Change the \code{seed_form} value of a \link{BPCellsMatrix} object.
}
}
\description{
\code{BPCellsArray} offers two seed form for using BPCells Matrix as the
DelayedArray seed:
\itemize{
\item \code{BPCells}: use the \code{IterableMatrix} of BPCells as the seed directly.
\item \code{DelayedArray}: convert \code{IterableMatrix} into a parallel
\link[DelayedArray:DelayedOp-class]{DelayedOp} object (\code{BPCellsDelayedOp}).
}

Both methods are generally effective for most operations. However, choosing
\code{seed_form=DelayedArray} ensures better compatibility with DelayedArray
package. It should be noted that \link[DelayedArray:showtree]{showtree} and
\link[DelayedArray:showtree]{seedApply} from DelayedArray do not function well
when \code{seed_form=BPCells}. Nonetheless, BPCellsArray provides its own
\link{seedApply} and \link{showtree} functions for both methods. Overall, executing
operations with \code{seed_form=BPCells} is expected to be faster due to the
absence of the need to convert between \code{IterableMatrix} and
\code{BPCellsDelayedOp} objects. \code{seed_form=BPCells} is also the default method.
}
