% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-showtree.R
\name{showtree}
\alias{showtree}
\alias{nseed,BPCellsMatrix-method}
\alias{seed,BPCellsMatrix-method}
\alias{path,BPCellsMatrix-method}
\alias{path,BPCellsBindMatrixSeed-method}
\alias{seed,BPCellsBindMatrixSeed-method}
\alias{nseed,BPCellsBindMatrixSeed-method}
\alias{path,BPCellsMaskSeed-method}
\alias{seed,BPCellsMaskSeed-method}
\alias{nseed,BPCellsMaskSeed-method}
\alias{path,BPCellsMultiplySeed-method}
\alias{seed,BPCellsMultiplySeed-method}
\alias{nseed,BPCellsMultiplySeed-method}
\alias{path,BPCellsUnaryOpsSeed-method}
\alias{seed,BPCellsUnaryOpsSeed-method}
\alias{nseed,BPCellsUnaryOpsSeed-method}
\alias{path,BPCellsMemSeed-method}
\alias{seed,BPCellsMemSeed-method}
\alias{nseed,BPCellsMemSeed-method}
\alias{path,BPCellsdgCMatrixSeed-method}
\alias{seed,BPCellsdgCMatrixSeed-method}
\alias{nseed,BPCellsdgCMatrixSeed-method}
\alias{path,BPCellsDirSeed-method}
\alias{seed,BPCellsDirSeed-method}
\alias{nseed,BPCellsDirSeed-method}
\title{Visualize and access the leaves of a tree of delayed operations}
\usage{
showtree(object)

\S4method{nseed}{BPCellsMatrix}(x)

\S4method{seed}{BPCellsMatrix}(x)

\S4method{path}{BPCellsMatrix}(object)

\S4method{path}{BPCellsBindMatrixSeed}(object)

\S4method{seed}{BPCellsBindMatrixSeed}(x)

\S4method{nseed}{BPCellsBindMatrixSeed}(x)

\S4method{path}{BPCellsMaskSeed}(object)

\S4method{seed}{BPCellsMaskSeed}(x)

\S4method{nseed}{BPCellsMaskSeed}(x)

\S4method{path}{BPCellsMultiplySeed}(object)

\S4method{seed}{BPCellsMultiplySeed}(x)

\S4method{nseed}{BPCellsMultiplySeed}(x)

\S4method{path}{BPCellsUnaryOpsSeed}(object)

\S4method{seed}{BPCellsUnaryOpsSeed}(x)

\S4method{nseed}{BPCellsUnaryOpsSeed}(x)

\S4method{path}{BPCellsMemSeed}(object)

\S4method{seed}{BPCellsMemSeed}(x)

\S4method{nseed}{BPCellsMemSeed}(x)

\S4method{path}{BPCellsdgCMatrixSeed}(object)

\S4method{seed}{BPCellsdgCMatrixSeed}(x)

\S4method{nseed}{BPCellsdgCMatrixSeed}(x)

\S4method{path}{BPCellsDirSeed}(object)

\S4method{seed}{BPCellsDirSeed}(x)

\S4method{nseed}{BPCellsDirSeed}(x)
}
\arguments{
\item{x, object}{A \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object}
}
\value{
\itemize{
\item \code{showtree}: return the input invisiblely
\item \code{path}: An atomic characters
\item \code{nseed}: An integer number
\item \code{seed}: A list of \link[=BPCellsSeed-class]{BPCellsSeed} objects or a single
\link[=BPCellsSeed-class]{BPCellsSeed} object
}
}
\description{
\code{showtree} can be used to visualize the tree of delayed operations carried by
a \code{BPCellsArray} object.
Use \code{nseed}, \code{seed}, or \code{path} to access the number of seeds, the seed, or
the seed path of a \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object, respectively.
}
