% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSVD.R
\docType{class}
\name{SpectraParam}
\alias{SpectraParam}
\alias{runSVD,SpectraParam-method}
\title{BiocSingularParam classes}
\usage{
SpectraParam()

\S4method{runSVD}{SpectraParam}(
  x,
  k,
  nu = k,
  nv = k,
  center = FALSE,
  scale = FALSE,
  ncv = NULL,
  tol = 1e-10,
  maxitr = 1000,
  threads = 0L,
  ...,
  BSPARAM
)
}
\arguments{
\item{x}{A numeric matrix-like object to use in the SVD.}

\item{k}{Number of singular values requested.}

\item{nu}{Number of right singular vectors to be computed. This must be between 0 and 'k'. (Must be equal to 'k' for BPCells IterableMatrix)}

\item{nv}{Number of right singular vectors to be computed. This must
be between 0 and \code{k}.}

\item{center}{Either a logical value (\code{TRUE}/\code{FALSE}), or a numeric
vector of length \eqn{n}. If a vector \eqn{c} is supplied, then SVD is
computed on the matrix \eqn{A - 1c'}{A - 1 * c'}, in an implicit way without
actually forming this matrix.  \code{center = TRUE} has the same effect as
\code{center = colMeans(A)}. Default is \code{FALSE}. Ignored if x is a
\code{IterableMatrix} object.}

\item{scale}{Either a logical value (\code{TRUE}/\code{FALSE}), or a numeric
vector of length \eqn{n}. If a vector \eqn{s} is supplied, then SVD is
computed on the matrix \eqn{(A - 1c')S}{(A - 1 * c')S}, where \eqn{c} is the
centering vector and \eqn{S = diag(1/s)}.If \code{scale = TRUE}, then the
vector \eqn{s} is computed as the column norm of \eqn{A - 1c'}{A - 1 *
c'}.Default is \code{FALSE}. Ignored if x is a \code{IterableMatrix} object.}

\item{ncv}{Number of Lanzcos basis vectors to use. More vectors will result
in faster convergence, but with greater memory use. \code{ncv} must be
satisfy \eqn{k < ncv \le p}{k < ncv <= p} where \code{p = min(m, n)}.
Default is \code{min(p, max(2*k+1, 20))}.}

\item{tol}{Precision parameter. Default is 1e-10.}

\item{maxitr}{Maximum number of iterations. Default is 1000.}

\item{threads}{Control threads to use calculating mat-vec producs (BPCells specific)}

\item{...}{Not used currently}

\item{BSPARAM}{A \linkS4class{BiocSingularParam} object specifying the type of algorithm to run.}
}
\description{
Find the Largest k Singular Values/Vectors of a Matrix using \code{RSpectra}
package.
}
\seealso{
\link[RSpectra:svds]{RSpectra} and \link[BPCells:svds]{BPCells}
}
