% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_storage.R
\name{transpose_storage}
\alias{transpose_storage}
\alias{transpose_storage,BPCellsSeed-method}
\alias{transpose_storage,BPCellsMatrix-method}
\title{Transpose the storage order for a \code{BPCellsSeed} or \code{BPCellsMatrix} object}
\usage{
transpose_storage(object, ...)

\S4method{transpose_storage}{BPCellsSeed}(object, ...)

\S4method{transpose_storage}{BPCellsMatrix}(object, ...)
}
\arguments{
\item{object}{A \link{BPCellsSeed} or \link{BPCellsMatrix} object.}

\item{...}{
  Arguments passed on to \code{\link[BPCells:transpose_storage_order]{BPCells::transpose_storage_order}}
  \describe{
    \item{\code{outdir}}{Directory to store the output}
    \item{\code{tmpdir}}{Temporary directory to use for intermediate storage}
    \item{\code{load_bytes}}{The minimum contiguous load size during the merge sort passes}
    \item{\code{sort_bytes}}{The amount of memory to allocate for re-sorting chunks of entries}
  }}
}
\value{
A \link{BPCellsSeed} or \link{BPCellsMatrix} object.
}
\description{
Transpose the storage order for a \code{BPCellsSeed} or \code{BPCellsMatrix} object
}
