% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MatrixDir.R
\name{writeBPCellsDirArray}
\alias{writeBPCellsDirArray}
\alias{writeBPCellsDirArray,ANY-method}
\alias{writeBPCellsDirArray,dgCMatrix-method}
\alias{writeBPCellsDirArray,IterableMatrix-method}
\title{Write a sparce matrices into a BPCells Directory of files format}
\usage{
writeBPCellsDirArray(x, ...)

\S4method{writeBPCellsDirArray}{ANY}(x, ...)

\S4method{writeBPCellsDirArray}{dgCMatrix}(
  x,
  path = NULL,
  compress = TRUE,
  buffer_size = 8192L,
  overwrite = FALSE
)

\S4method{writeBPCellsDirArray}{IterableMatrix}(
  x,
  path = NULL,
  compress = TRUE,
  buffer_size = 8192L,
  overwrite = FALSE
)
}
\arguments{
\item{x}{Input matrix, any matrix can be coerced into
\link[Matrix:dgCMatrix-class]{dgCMatrix} object.}

\item{...}{Additional parameters passed to specific methods.}

\item{path}{Directory to save the data into, if \code{NULL}, will use a temporary
directory.}

\item{compress}{Whether or not to compress the data.}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}

\item{overwrite}{If \code{TRUE}, write to a temp dir then overwrite existing data. Alternatively,
pass a temp path as a string to customize the temp dir location.}
}
\value{
A \link{BPCellsDir} object.
}
\description{
Write a sparce matrices into a BPCells Directory of files format
}
