% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-Arith.R
\name{BPCells-Arithmetic}
\alias{BPCells-Arithmetic}
\alias{BPCells-Arith}
\alias{Arith,BPCellsMatrix,numeric-method}
\alias{Arith,numeric,BPCellsMatrix-method}
\alias{pow_slow}
\alias{pow_slow,BPCellsMatrix-method}
\title{Arithmetic operators for BPCellsMatrix}
\usage{
\S4method{Arith}{BPCellsMatrix,numeric}(e1, e2)

\S4method{Arith}{numeric,BPCellsMatrix}(e1, e2)

pow_slow(e1, e2)

\S4method{pow_slow}{BPCellsMatrix}(e1, e2)
}
\arguments{
\item{e1, e2}{One of \code{x} or \code{y} must be \link[=BPCellsMatrix-class]{BPCellsMatrix}
object, and the another can be a \link[=BPCellsMatrix-class]{BPCellsMatrix} object
or any objects can be converted into \link{BPCellsSeed} object.}
}
\value{
A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
Arithmetic operators for BPCellsMatrix
}
\section{Arithmetic operators}{

\itemize{
\item \code{BPCells}: \code{+}, \code{-}, \code{*}, \code{/}, \code{^}
\item \code{DelayedArray}: \code{\%\%}, - \code{\%/\%}
}
}

\seealso{
\link{BPCellsMatrix}
}
