% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-Arith.R
\name{BPCells-Arithmetic}
\alias{BPCells-Arithmetic}
\alias{BPCells-Arith}
\alias{Arith,BPCellsMatrix,numeric-method}
\alias{Arith,numeric,BPCellsMatrix-method}
\alias{pow_slow}
\alias{pow_slow,BPCellsMatrix-method}
\title{Arithmetic operators for BPCellsMatrix}
\usage{
\S4method{Arith}{BPCellsMatrix,numeric}(e1, e2)

\S4method{Arith}{numeric,BPCellsMatrix}(e1, e2)

pow_slow(e1, e2)

\S4method{pow_slow}{BPCellsMatrix}(e1, e2)
}
\value{
A \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object depends on the class of \code{e1} /
\code{e2}.
}
\description{
Arithmetic operators for BPCellsMatrix
}
