% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-Arithmetic.R
\name{BPCells-Arithmetic}
\alias{BPCells-Arithmetic}
\alias{*,BPCellsMatrix,numeric-method}
\alias{*,numeric,BPCellsMatrix-method}
\alias{+,BPCellsMatrix,numeric-method}
\alias{+,numeric,BPCellsMatrix-method}
\alias{/,BPCellsMatrix,numeric-method}
\alias{-,BPCellsMatrix,numeric-method}
\alias{-,numeric,BPCellsMatrix-method}
\title{Arithmetic operators for BPCellsMatrix}
\usage{
\S4method{*}{BPCellsMatrix,numeric}(e1, e2)

\S4method{*}{numeric,BPCellsMatrix}(e1, e2)

\S4method{+}{BPCellsMatrix,numeric}(e1, e2)

\S4method{+}{numeric,BPCellsMatrix}(e1, e2)

\S4method{/}{BPCellsMatrix,numeric}(e1, e2)

\S4method{-}{BPCellsMatrix,numeric}(e1, e2)

\S4method{-}{numeric,BPCellsMatrix}(e1, e2)
}
\arguments{
\item{e1, e2}{See detailed methods signature.}
}
\description{
Arithmetic operators for BPCellsMatrix
}
