% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-crossprod.R
\name{BPCells-crossprod}
\alias{BPCells-crossprod}
\alias{crossprod}
\alias{crossprod,BPCellsMatrix,BPCellsMatrix-method}
\alias{crossprod,BPCellsMatrix,dgCMatrix-method}
\alias{crossprod,dgCMatrix,BPCellsMatrix-method}
\alias{crossprod,BPCellsMatrix,ANY-method}
\alias{crossprod,ANY,BPCellsMatrix-method}
\alias{crossprod,BPCellsMatrix,matrix-method}
\alias{crossprod,matrix,BPCellsMatrix-method}
\title{Matrix Crossproduct}
\usage{
\S4method{crossprod}{BPCellsMatrix,BPCellsMatrix}(x, y)

\S4method{crossprod}{BPCellsMatrix,dgCMatrix}(x, y)

\S4method{crossprod}{dgCMatrix,BPCellsMatrix}(x, y)

\S4method{crossprod}{BPCellsMatrix,ANY}(x, y)

\S4method{crossprod}{ANY,BPCellsMatrix}(x, y)

\S4method{crossprod}{BPCellsMatrix,matrix}(x, y)

\S4method{crossprod}{matrix,BPCellsMatrix}(x, y)
}
\arguments{
\item{x, y}{A \link{BPCellsMatrix} object or matrx-like object which can be coerced
into a \link[Matrix:dgCMatrix-class]{dgCMatrix}.}
}
\value{
Matrix Crossproduct, a \link{BPCellsMatrix} object or a dense matrix
(matrix and numeric methods).
}
\description{
Given matrices x and y as arguments, return a matrix cross-product.
}
\seealso{
\link{\%*\%}
}
