% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-crossprod.R
\name{BPCells-crossprod}
\alias{BPCells-crossprod}
\alias{crossprod}
\alias{crossprod,BPCellsMatrix,BPCellsMatrix-method}
\alias{crossprod,BPCellsMatrix,dgCMatrix-method}
\alias{crossprod,dgCMatrix,BPCellsMatrix-method}
\alias{crossprod,BPCellsMatrix,ANY-method}
\alias{crossprod,ANY,BPCellsMatrix-method}
\alias{crossprod,BPCellsMatrix,matrix-method}
\alias{crossprod,matrix,BPCellsMatrix-method}
\title{Matrix Crossproduct}
\usage{
\S4method{crossprod}{BPCellsMatrix,BPCellsMatrix}(x, y)

\S4method{crossprod}{BPCellsMatrix,dgCMatrix}(x, y)

\S4method{crossprod}{dgCMatrix,BPCellsMatrix}(x, y)

\S4method{crossprod}{BPCellsMatrix,ANY}(x, y)

\S4method{crossprod}{ANY,BPCellsMatrix}(x, y)

\S4method{crossprod}{BPCellsMatrix,matrix}(x, y)

\S4method{crossprod}{matrix,BPCellsMatrix}(x, y)
}
\arguments{
\item{x, y}{One of \code{x} or \code{y} must be \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object, and the another must be a matrix
which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix}.}
}
\value{
A dense matrix if one of \code{x} or \code{y} is a regular matrix or atomic
vector. Otherwise, a \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object depends on the class of \code{x} or
\code{y}.
}
\description{
Given matrices x and y as arguments, return a matrix cross-product.
}
\seealso{
\itemize{
\item \link[=BPCells-Multiplication]{\%*\%}
\item \link[=BPCells-tcrossprod]{tcrossprod}
}
}
