% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-Multiply.R
\name{BPCells-Multiplication}
\alias{BPCells-Multiplication}
\alias{\%*\%,BPCellsMatrix,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,ANY-method}
\alias{\%*\%,ANY,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,matrix-method}
\alias{\%*\%,matrix,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,numeric-method}
\alias{\%*\%,numeric,BPCellsMatrix-method}
\title{Matrix Multiplication}
\usage{
\S4method{\%*\%}{BPCellsMatrix,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,ANY}(x, y)

\S4method{\%*\%}{ANY,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,matrix}(x, y)

\S4method{\%*\%}{matrix,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,numeric}(x, y)

\S4method{\%*\%}{numeric,BPCellsMatrix}(x, y)
}
\arguments{
\item{x, y}{One of \code{x} or \code{y} must be \link[=BPCellsMatrix-class]{BPCellsMatrix}
object, and the another must be a \link[=BPCellsMatrix-class]{BPCellsMatrix} object
or  any objects can be converted into \link{BPCellsSeed} object.}
}
\value{
A dense matrix if one of \code{x} or \code{y} is a regular matrix or atomic
vector. Otherwise, a \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
Multiplies two matrices, if they are conformable. If one argument is a
vector, it will be promoted to either a row or column matrix to make the two
arguments conformable. If both are vectors of the same length, it will return
the inner product (as a matrix).
}
\seealso{
\itemize{
\item \link[=BPCells-crossprod]{crossprod}
\item \link[=BPCells-tcrossprod]{tcrossprod}
}
}
