% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-multiplication.R
\name{BPCells-Multiplication}
\alias{BPCells-Multiplication}
\alias{\%*\%,BPCellsMatrix,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,dgCMatrix-method}
\alias{\%*\%,dgCMatrix,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,ANY-method}
\alias{\%*\%,ANY,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,matrix-method}
\alias{\%*\%,matrix,BPCellsMatrix-method}
\alias{\%*\%,BPCellsMatrix,numeric-method}
\alias{\%*\%,numeric,BPCellsMatrix-method}
\title{Matrix Multiplication}
\usage{
\S4method{\%*\%}{BPCellsMatrix,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,dgCMatrix}(x, y)

\S4method{\%*\%}{dgCMatrix,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,ANY}(x, y)

\S4method{\%*\%}{ANY,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,matrix}(x, y)

\S4method{\%*\%}{matrix,BPCellsMatrix}(x, y)

\S4method{\%*\%}{BPCellsMatrix,numeric}(x, y)

\S4method{\%*\%}{numeric,BPCellsMatrix}(x, y)
}
\arguments{
\item{x, y}{A \link{BPCellsMatrix} object or matrx-like object which can be coerced
into a \link[Matrix:dgCMatrix-class]{dgCMatrix}.}
}
\value{
\code{x \%*\% y}: Matrix multiplication, a \link{BPCellsMatrix} object or a dense
matrix (matrix and numeric methods).
}
\description{
Multiplies two matrices, if they are conformable. If one argument is a
vector, it will be promoted to either a row or column matrix to make the two
arguments conformable. If both are vectors of the same length, it will return
the inner product (as a matrix).
}
