% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-summarization.R
\name{BPCells-Summarization}
\alias{BPCells-Summarization}
\alias{matrix_stats}
\alias{matrix_stats,BPCellsSeed-method}
\alias{matrix_stats,BPCellsMatrix-method}
\alias{rowSums,BPCellsMatrix-method}
\alias{rowSums}
\alias{colSums,BPCellsMatrix-method}
\alias{colSums}
\alias{rowMeans,BPCellsMatrix-method}
\alias{rowMeans}
\alias{colMeans,BPCellsMatrix-method}
\alias{colMeans}
\alias{rowVars,BPCellsSeed-method}
\alias{rowVars}
\alias{colVars,BPCellsSeed-method}
\alias{colVars}
\alias{rowVars,BPCellsMatrix-method}
\alias{colVars,BPCellsMatrix-method}
\alias{rowSds,BPCellsSeed-method}
\alias{rowSds}
\alias{colSds,BPCellsSeed-method}
\alias{colSds}
\alias{rowSds,BPCellsMatrix-method}
\alias{colSds,BPCellsMatrix-method}
\title{BPCellsMatrix row/col summarization}
\usage{
matrix_stats(object, ...)

\S4method{matrix_stats}{BPCellsSeed}(object, ...)

\S4method{matrix_stats}{BPCellsMatrix}(object, ...)

\S4method{rowSums}{BPCellsMatrix}(x)

\S4method{colSums}{BPCellsMatrix}(x)

\S4method{rowMeans}{BPCellsMatrix}(x)

\S4method{colMeans}{BPCellsMatrix}(x)

\S4method{rowVars}{BPCellsSeed}(x)

\S4method{colVars}{BPCellsSeed}(x)

\S4method{rowVars}{BPCellsMatrix}(x)

\S4method{colVars}{BPCellsMatrix}(x)

\S4method{rowSds}{BPCellsSeed}(x)

\S4method{colSds}{BPCellsSeed}(x)

\S4method{rowSds}{BPCellsMatrix}(x)

\S4method{colSds}{BPCellsMatrix}(x)
}
\arguments{
\item{object}{A \link{BPCellsMatrix} object.}

\item{...}{
  Arguments passed on to \code{\link[BPCells:matrix_stats]{BPCells::matrix_stats}}
  \describe{
    \item{\code{row_stats}}{Which row statistics to compute}
    \item{\code{col_stats}}{Which col statistics to compute}
    \item{\code{threads}}{Number of threads to use during execution}
  }}

\item{x}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} or
\link[=BPCellsSeed-class]{BPcellsSeed}object.}
}
\value{
\itemize{
\item \code{matrix_stats}: A list of
\itemize{
\item \code{row_stats}: matrix of n_stats x n_rows
\item \code{col_stats}: matrix of n_stats x n_cols
}
}

\itemize{
\item \code{rowSums()}: vector of row sums
}

\itemize{
\item \code{colSums()}: vector of column sums
}

\itemize{
\item \code{rowMeans()}: vector of row means
}

\itemize{
\item \code{colMeans()}: vector of column means
}

\itemize{
\item \code{rowVars()}: vector of row variance
}

\itemize{
\item \code{colVars()}: vector of column variance
}

\itemize{
\item \code{rowSds()}: vector of row Standard Deviation
}

\itemize{
\item \code{colSds()}: vector of column Standard Deviation
}
}
\description{
Calculate matrix stats
}
\details{
The statistics will be calculated in a single pass over the matrix,
so this method is desirable to use for efficiency purposes compared to
the more standard rowMeans or colMeans if multiple statistics are needed.
The stats are ordered by complexity: nonzero, mean, then variance. All
less complex stats are calculated in the process of calculating a more complicated stat.
So to calculate mean and variance simultaneously, just ask for variance,
which will compute mean and nonzero counts as a side-effect
}
