% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-summarization.R
\name{BPCells-Summarization}
\alias{BPCells-Summarization}
\alias{rowSums,BPCellsMatrix-method}
\alias{rowSums}
\alias{colSums,BPCellsMatrix-method}
\alias{colSums}
\alias{rowMeans,BPCellsMatrix-method}
\alias{rowMeans}
\alias{colMeans,BPCellsMatrix-method}
\alias{colMeans}
\alias{rowVars,BPCellsSeed-method}
\alias{rowVars}
\alias{colVars,BPCellsSeed-method}
\alias{colVars}
\alias{rowVars,BPCellsMatrix-method}
\alias{colVars,BPCellsMatrix-method}
\alias{rowSds,BPCellsSeed-method}
\alias{rowSds}
\alias{colSds,BPCellsSeed-method}
\alias{colSds}
\alias{rowSds,BPCellsMatrix-method}
\alias{colSds,BPCellsMatrix-method}
\title{BPCellsMatrix row/col summarization}
\usage{
\S4method{rowSums}{BPCellsMatrix}(x)

\S4method{colSums}{BPCellsMatrix}(x)

\S4method{rowMeans}{BPCellsMatrix}(x)

\S4method{colMeans}{BPCellsMatrix}(x)

\S4method{rowVars}{BPCellsSeed}(x)

\S4method{colVars}{BPCellsSeed}(x)

\S4method{rowVars}{BPCellsMatrix}(x)

\S4method{colVars}{BPCellsMatrix}(x)

\S4method{rowSds}{BPCellsSeed}(x)

\S4method{colSds}{BPCellsSeed}(x)

\S4method{rowSds}{BPCellsMatrix}(x)

\S4method{colSds}{BPCellsMatrix}(x)
}
\arguments{
\item{x}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} or
\link[=BPCellsSeed-class]{BPcellsSeed}object.}
}
\value{
\itemize{
\item \code{rowSums()}: vector of row sums
}

\itemize{
\item \code{colSums()}: vector of column sums
}

\itemize{
\item \code{rowMeans()}: vector of row means
}

\itemize{
\item \code{colMeans()}: vector of column means
}

\itemize{
\item \code{rowVars()}: vector of row variance
}

\itemize{
\item \code{colVars()}: vector of column variance
}

\itemize{
\item \code{rowSds()}: vector of row Standard Deviation
}

\itemize{
\item \code{colSds()}: vector of column Standard Deviation
}
}
\description{
BPCellsMatrix row/col summarization
}
