% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-tcrossprod.R
\name{BPCells-tcrossprod}
\alias{BPCells-tcrossprod}
\alias{tcrossprod}
\alias{tcrossprod,BPCellsMatrix,BPCellsMatrix-method}
\alias{tcrossprod,BPCellsMatrix,dgCMatrix-method}
\alias{tcrossprod,dgCMatrix,BPCellsMatrix-method}
\alias{tcrossprod,BPCellsMatrix,ANY-method}
\alias{tcrossprod,ANY,BPCellsMatrix-method}
\alias{tcrossprod,BPCellsMatrix,matrix-method}
\alias{tcrossprod,matrix,BPCellsMatrix-method}
\alias{tcrossprod,BPCellsSeed,BPCellsSeed-method}
\alias{tcrossprod,BPCellsSeed,dgCMatrix-method}
\alias{tcrossprod,dgCMatrix,BPCellsSeed-method}
\alias{tcrossprod,BPCellsSeed,matrix-method}
\alias{tcrossprod,matrix,BPCellsSeed-method}
\alias{tcrossprod,BPCellsSeed,numeric-method}
\alias{tcrossprod,numeric,BPCellsSeed-method}
\title{Matrix Products of Transpose}
\usage{
\S4method{tcrossprod}{BPCellsMatrix,BPCellsMatrix}(x, y)

\S4method{tcrossprod}{BPCellsMatrix,dgCMatrix}(x, y)

\S4method{tcrossprod}{dgCMatrix,BPCellsMatrix}(x, y)

\S4method{tcrossprod}{BPCellsMatrix,ANY}(x, y)

\S4method{tcrossprod}{ANY,BPCellsMatrix}(x, y)

\S4method{tcrossprod}{BPCellsMatrix,matrix}(x, y)

\S4method{tcrossprod}{matrix,BPCellsMatrix}(x, y)

\S4method{tcrossprod}{BPCellsSeed,BPCellsSeed}(x, y)

\S4method{tcrossprod}{BPCellsSeed,dgCMatrix}(x, y)

\S4method{tcrossprod}{dgCMatrix,BPCellsSeed}(x, y)

\S4method{tcrossprod}{BPCellsSeed,matrix}(x, y)

\S4method{tcrossprod}{matrix,BPCellsSeed}(x, y)

\S4method{tcrossprod}{BPCellsSeed,numeric}(x, y)

\S4method{tcrossprod}{numeric,BPCellsSeed}(x, y)
}
\arguments{
\item{x, y}{One of \code{x} or \code{y} must be \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object, and the another must be a matrix
which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix}.}
}
\value{
A dense matrix if one of \code{x} or \code{y} is a regular matrix or atomic
vector. Otherwise, a \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object depends on the class of \code{x} or
\code{y}.
}
\description{
Given matrices x and y as arguments, return a matrix product of Transpose.
}
\seealso{
\itemize{
\item \link[=BPCells-Multiplication]{\%*\%}
\item \link[=BPCells-crossprod]{crossprod}
}
}
