% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Convert.R
\docType{class}
\name{BPCellsConvert}
\alias{BPCellsConvert}
\alias{BPCellsConvertArray-class}
\alias{DelayedArray,BPCellsConvertSeed-method}
\alias{BPCellsConvertArray}
\alias{BPCellsConvertMatrix-class}
\alias{matrixClass,BPCellsConvertArray-method}
\title{Delayed BPCells ConvertMatrixType}
\usage{
\S4method{DelayedArray}{BPCellsConvertSeed}(seed)

BPCellsConvertArray(x)

\S4method{matrixClass}{BPCellsConvertArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsConvertSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsConvertArray}: A \code{ConvertMatrixType} object.
\item \code{matrixClass}: A \code{BPCellsConvertArray} object.
}}
}
\description{
The \code{BPCellsConvertArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{ConvertMatrixType}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link{convert_type} to create a \code{BPCellsConvert} object.
}
\seealso{
\link{BPCellsSeed}
}
