% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Dir.R
\docType{class}
\name{BPCellsDir}
\alias{BPCellsDir}
\alias{BPCellsDirSeed}
\alias{BPCellsDirArray-class}
\alias{DelayedArray,BPCellsDirSeed-method}
\alias{BPCellsDirArray}
\alias{BPCellsDirMatrix-class}
\alias{matrixClass,BPCellsDirArray-method}
\alias{path,BPCellsDirSeed-method}
\alias{[,BPCellsDirSeed,ANY,ANY,ANY-method}
\title{Delayed MatrixDir arrays}
\usage{
BPCellsDirSeed(x, buffer_size = 8192L)

\S4method{DelayedArray}{BPCellsDirSeed}(seed)

BPCellsDirArray(x, ...)

\S4method{matrixClass}{BPCellsDirArray}(x)

\S4method{path}{BPCellsDirSeed}(object)

\S4method{[}{BPCellsDirSeed,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{BPCellsDirSeed} or \code{BPCellsDirArray} object. For following
functions:
\itemize{
\item \code{BPCellsDirSeed} and \code{BPCellsDirArray}: A string for the path of the
BPCells matrix dir.
}}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}

\item{seed}{A \code{BPCellsDirSeed} object.}

\item{...}{\itemize{
\item \code{BPCellsDirArray}: Additional parameters passed into \code{BPCellsDirSeed}.
\item \code{[}: Not used currently.
}}

\item{object}{A \code{BPCellsMatrix} object.}

\item{i, j}{Row and Column index.}

\item{drop}{Ignored, always be \code{FALSE}.}
}
\description{
The \code{BPCellsDirArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixDir} object in
BPCells.
}
\seealso{
\link{writeBPCellsDirArray}
}
