% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MatrixDir.R
\docType{class}
\name{BPCellsDir}
\alias{BPCellsDir}
\alias{BPCellsDirSeed}
\alias{BPCellsDirArray-class}
\alias{DelayedArray,BPCellsDirSeed-method}
\alias{BPCellsDirArray}
\alias{BPCellsDirMatrix-class}
\alias{matrixClass,BPCellsDirArray-method}
\alias{show,BPCellsDirSeed-method}
\alias{path,BPCellsDirSeed-method}
\alias{type,BPCellsDirSeed-method}
\alias{is_sparse,BPCellsDirSeed-method}
\alias{extract_array,BPCellsDirSeed-method}
\alias{extract_sparse_array,BPCellsDirSeed-method}
\alias{[,BPCellsDirSeed,ANY,ANY,ANY-method}
\alias{[,BPCellsDirArray,ANY,ANY,ANY-method}
\title{Delayed MatrixDir arrays}
\usage{
BPCellsDirSeed(x, buffer_size = 8192L)

\S4method{DelayedArray}{BPCellsDirSeed}(seed)

BPCellsDirArray(x, ...)

\S4method{matrixClass}{BPCellsDirArray}(x)

\S4method{show}{BPCellsDirSeed}(object)

\S4method{path}{BPCellsDirSeed}(object)

\S4method{type}{BPCellsDirSeed}(x)

\S4method{is_sparse}{BPCellsDirSeed}(x)

\S4method{extract_array}{BPCellsDirSeed}(x, index)

\S4method{extract_sparse_array}{BPCellsDirSeed}(x, index)

\S4method{[}{BPCellsDirSeed,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{BPCellsDirArray,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{For \code{BPCellsDirSeed} and \code{BPCellsDirArray}, a path of the
\code{MatrixDir} data, or a \code{MatrixDir} object. For other function, a
\code{BPCellsDirSeed} or \code{BPCellsDirArray} object.}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}

\item{seed}{A \code{BPCellsDirSeed} object.}

\item{...}{Additional parameters passed into \code{BPCellsDirSeed}.}

\item{object}{A \code{BPCellsDirSeed} object.}

\item{index}{
    An unnamed list of integer vectors, one per dimension in \code{x}.
    Each vector is called a \emph{subscript} and can only contain
    positive integers that are valid 1-based indices along the corresponding
    dimension in \code{x}.

    Empty or missing subscripts are allowed. They must be represented
    by list elements set to \code{integer(0)} or \code{NULL}, respectively.

    The subscripts cannot contain NAs or non-positive values.

    Individual subscripts are allowed to contain duplicated indices.
  }

\item{i, j}{Row and Column index.}

\item{drop}{Not used, always be \code{FALSE}.}
}
\description{
The \code{BPCellsDirArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixDir} object in
BPCells.
}
\seealso{
\link{writeBPCellsDirArray}
}
