% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Dir.R, R/Class-Multiply.R
\docType{class}
\name{BPCellsDir}
\alias{BPCellsDir}
\alias{BPCellsDirSeed}
\alias{BPCellsDirArray-class}
\alias{DelayedArray,BPCellsDirSeed-method}
\alias{BPCellsDirArray}
\alias{BPCellsDirMatrix-class}
\alias{matrixClass,BPCellsDirArray-method}
\alias{path,BPCellsDirSeed-method}
\alias{[,BPCellsDirSeed,ANY,ANY,ANY-method}
\title{Delayed MatrixDir arrays}
\usage{
BPCellsDirSeed(x, buffer_size = 8192L)

\S4method{DelayedArray}{BPCellsDirSeed}(seed)

BPCellsDirArray(x, ...)

\S4method{matrixClass}{BPCellsDirArray}(x)

\S4method{path}{BPCellsDirSeed}(object)

\S4method{[}{BPCellsDirSeed,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{BPCellsDirSeed,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{For \code{BPCellsDirSeed} and \code{BPCellsDirArray}, a path of the
\code{MatrixDir} data, or a \code{MatrixDir} object. For other function, a
\code{BPCellsDirSeed} or \code{BPCellsDirArray} object.}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}

\item{seed}{A \code{BPCellsDirSeed} object.}

\item{...}{Additional parameters passed into \code{BPCellsDirSeed}.}

\item{i, j}{Row and Column index.}

\item{drop}{Not used, always be \code{FALSE}.}
}
\description{
The \code{BPCellsDirArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixDir} object in
BPCells.
}
\seealso{
\link{writeBPCellsDirArray}
}
