% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Mask.R
\docType{class}
\name{BPCellsMask}
\alias{BPCellsMask}
\alias{BPCellsMaskArray-class}
\alias{DelayedArray,BPCellsMaskSeed-method}
\alias{BPCellsMaskArray}
\alias{BPCellsMaskMatrix-class}
\alias{matrixClass,BPCellsMaskArray-method}
\title{Delayed BPCells MatrixMask}
\usage{
\S4method{DelayedArray}{BPCellsMaskSeed}(seed)

BPCellsMaskArray(x)

\S4method{matrixClass}{BPCellsMaskArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsMaskSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsMatrixMaskArray}: A \code{MatrixMask} object.
\item \code{matrixClass}: A \code{BPCellsMatrixMaskArray} object.
}}
}
\description{
The \code{BPCellsMaskArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixMask}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link{mask_matrix} of other BPCellsMatrix objects to create a \code{BPCellsMaskMatrix}.
}
\seealso{
\link{BPCellsSeed}
}
