% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BPCellsMatrix.R, R/Class-subset.R,
%   R/Method-Subassign.R, R/Class-RenameDims.R
\docType{class}
\name{show,BPCellsArray-method}
\alias{show,BPCellsArray-method}
\alias{[,BPCellsMatrix,ANY,ANY,ANY-method}
\alias{[<-,BPCellsMatrix,ANY,ANY,ANY-method}
\alias{BPCellsMatrix-class}
\alias{BPCellsMatrix-methods}
\alias{BPCellsArray}
\alias{BPCellsArray-class}
\alias{matrixClass,BPCellsArray-method}
\alias{DelayedArray,IterableMatrix-method}
\alias{as.array.BPCellsMatrix}
\alias{as.array,BPCellsMatrix-method}
\alias{as.matrix.BPCellsMatrix}
\alias{as.matrix,BPCellsMatrix-method}
\alias{t,BPCellsMatrix-method}
\alias{t}
\alias{dimnames<-,BPCellsMatrix,ListOrNULL-method}
\alias{dimnames<-}
\alias{rownames<-,BPCellsMatrix,atomic-method}
\alias{rownames<-}
\alias{colnames<-,BPCellsMatrix,atomic-method}
\alias{colnames<-}
\title{DelayedArray backend of BPCells matrix}
\usage{
\S4method{show}{BPCellsArray}(object)

\S4method{[}{BPCellsMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{BPCellsMatrix,ANY,ANY,ANY}(x, i, j, ...) <- value

BPCellsArray(x)

\S4method{matrixClass}{BPCellsArray}(x)

\S4method{DelayedArray}{IterableMatrix}(seed)

\method{as.array}{BPCellsMatrix}(x, drop = FALSE)

\S4method{as.array}{BPCellsMatrix}(x, drop = FALSE)

\method{as.matrix}{BPCellsMatrix}(x)

\S4method{as.matrix}{BPCellsMatrix}(x)

\S4method{t}{BPCellsMatrix}(x)

\S4method{dimnames}{BPCellsMatrix,ListOrNULL}(x) <- value

\S4method{rownames}{BPCellsMatrix,atomic}(x) <- value

\S4method{colnames}{BPCellsMatrix,atomic}(x) <- value
}
\arguments{
\item{x}{\itemize{
\item \code{BPCellsArray} and \code{BPCellsMatrix}: Details see \link{BPCellsSeed} for
supported object.
\item \code{matrixClass}: A string.
}}

\item{seed}{A \code{IterableMatrix} or \code{BPCellsDelayedOp} object.}
}
\value{
\itemize{
\item \code{[}: A \link{BPCellsDelayedSubset} object.
}

\itemize{
\item \verb{[<-}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \code{t}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \verb{dimnames<-}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \verb{rownames<-}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \verb{colnames<-}: A \link{BPCellsMatrix} object.
}
}
\description{
The \code{BPCellsMatrix} class just inherits from the
\link[DelayedArray:DelayedArray-class]{DelayedMatrix} class.
}
\section{Slots}{

\describe{
\item{\code{seed}}{A \link[=BPCellsSeed-class]{BPCellsSeed} object.}
}}

