% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BPCellsMatrix.R, R/Class-RenameDims.R,
%   R/Class-Subset.R, R/Method-Subassign.R
\docType{class}
\name{show,BPCellsArray-method}
\alias{show,BPCellsArray-method}
\alias{show,BPCellsMatrix-method}
\alias{BPCellsMatrix-class}
\alias{BPCellsMatrix-methods}
\alias{BPCellsArray}
\alias{BPCellsMatrix}
\alias{BPCellsArray-class}
\alias{matrixClass,BPCellsArray-method}
\alias{DelayedArray,IterableMatrix-method}
\alias{DelayedArray,BPCellsDelayedOp-method}
\alias{as.array.BPCellsMatrix}
\alias{as.array,BPCellsMatrix-method}
\alias{as.matrix.BPCellsMatrix}
\alias{as.matrix,BPCellsMatrix-method}
\alias{t,BPCellsMatrix-method}
\alias{t}
\alias{dimnames<-,BPCellsMatrix,ListOrNULL-method}
\alias{dimnames<-}
\alias{rownames<-,BPCellsMatrix-method}
\alias{rownames<-}
\alias{colnames<-,BPCellsMatrix-method}
\alias{colnames<-}
\alias{[,BPCellsMatrix,ANY,ANY,ANY-method}
\alias{[<-,BPCellsMatrix,ANY,ANY,ANY-method}
\title{DelayedArray backend of BPCells matrix}
\usage{
\S4method{show}{BPCellsArray}(object)

\S4method{show}{BPCellsMatrix}(object)

BPCellsArray(x, delayed = NULL)

BPCellsMatrix(x, delayed = NULL)

\S4method{matrixClass}{BPCellsArray}(x)

\S4method{DelayedArray}{IterableMatrix}(seed)

\S4method{DelayedArray}{BPCellsDelayedOp}(seed)

\method{as.array}{BPCellsMatrix}(x, drop = FALSE)

\S4method{as.array}{BPCellsMatrix}(x, drop = FALSE)

\method{as.matrix}{BPCellsMatrix}(x)

\S4method{as.matrix}{BPCellsMatrix}(x)

\S4method{t}{BPCellsMatrix}(x)

\S4method{dimnames}{BPCellsMatrix,ListOrNULL}(x) <- value

\S4method{rownames}{BPCellsMatrix}(x) <- value

\S4method{colnames}{BPCellsMatrix}(x) <- value

\S4method{[}{BPCellsMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{BPCellsMatrix,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x, object}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object}

\item{delayed}{A bool value. If \code{TRUE}, will convert \code{IterableMatrix} into a
parallel \link[DelayedArray:DelayedOp-class]{DelayedOp} object
(\code{BPCellsDelayedOp}), if \code{FALSE}, will use the \code{IterableMatrix} object as
the seed directly. see \link{set_delayed} for details.}

\item{seed}{A \link[=BPCellsSeed-class]{IterableMatrix} or
\link[=BPCellsSeed-class]{BPCellsDelayedOp} object.}

\item{drop}{A bool, if \code{TRUE}, any extents of length one will be removed and
return an atomic vector.}

\item{value}{\itemize{
\item \verb{dimnames<-}: A list of dimnames or \code{NULL}.
\item \verb{[<-}: A \link[=BPCellsMatrix-class]{BPCellsMatrix} object or any objects can be
converted into \link{BPCellsSeed} object.
}}

\item{i, j}{Row and Column index.}

\item{...}{Not used currently.}
}
\value{
\itemize{
\item \code{BPCellsArray} and \code{BPCellsMatrix}: A \code{BPCellsMatrix} object, since
\code{BPCells} can only support 2-dim array.
}

\itemize{
\item \code{matrixClass}: A string, always be \code{"BPCellsMatrix"}.
}

\itemize{
\item \code{as.array}: A dense matrix or an atomic vector.
}

\itemize{
\item \code{as.matrix}: A dense matrix.
}

\itemize{
\item \code{t}: A \code{BPCellsMatrix} object.
}

\itemize{
\item \verb{dimnames<-}: A \code{BPCellsMatrix} object.
}

\itemize{
\item \verb{rownames<-}: A \code{BPCellsMatrix} object.
}

\itemize{
\item \verb{colnames<-}: A \code{BPCellsMatrix} object.
}

\itemize{
\item \code{[}: A \code{BPCellsMatrix} object or an atomic vector.
}

\itemize{
\item \verb{[<-}: A \code{BPCellsMatrix} object.
}
}
\description{
The \code{BPCellsMatrix} class just inherits from the
\link[DelayedArray:DelayedArray-class]{DelayedMatrix} class.
}
\seealso{
\itemize{
\item \link[=BPCells-bind]{bind}: Combine two Objects by Columns or Rows.
\item \link[=BPCells-Multiplication]{\%*\%}: Matrix Multiplication.
\item \link[=BPCells-crossprod]{crossprod}: Matrix Crossproduct.
\item \link[=BPCells-Summarization]{summarization}: row/col summarization.
\item \link[=BPCells-Arithmetic]{Arith}: Binary Arithmetic operators.
\item \link[=BPCells-Math]{Math}: Math operators.
\item \link[=BPCells-Compare]{Compare}: Compare matrix.
\item \link[=pmin2]{pmin2/pmax2}: Maxima and Minima.
\item \link{DelayedArray-utils}: Common operations on DelayedArray objects
}
}
