% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BPCellsMatrix.R,
%   R/Transform-Round.R, R/Transform-expm1.R, R/Transform-expm1_slow.R,
%   R/Transform-log1p.R, R/Transform-log1p_slow.R, R/Transform-min_by_col.R,
%   R/Transform-min_by_row.R, R/Transform-min_scalar.R, R/Transform-pow.R,
%   R/Transform-pow_slow.R
\docType{class}
\name{BPCellsMatrix-class}
\alias{BPCellsMatrix-class}
\alias{show,BPCellsMatrix-method}
\alias{t,BPCellsMatrix-method}
\alias{[,BPCellsMatrix,ANY,ANY,ANY-method}
\alias{[<-,BPCellsMatrix,ANY,ANY,ANY-method}
\alias{dimnames<-,BPCellsMatrix,ListOrNULL-method}
\alias{round,BPCellsMatrix-method}
\alias{expm1,BPCellsMatrix-method}
\alias{expm1_slow,BPCellsMatrix-method}
\alias{log1p,BPCellsMatrix-method}
\alias{log1p_slow,BPCellsMatrix-method}
\alias{pmin_by_col,BPCellsMatrix-method}
\alias{pmin_by_row,BPCellsMatrix-method}
\alias{pmin_scalar,BPCellsMatrix-method}
\alias{^,BPCellsMatrix,ANY-method}
\alias{^,BPCellsMatrix,numeric-method}
\title{Base Class for Delayed BPCells matrix}
\usage{
\S4method{show}{BPCellsMatrix}(object)

\S4method{t}{BPCellsMatrix}(x)

\S4method{[}{BPCellsMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{BPCellsMatrix,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{dimnames}{BPCellsMatrix,ListOrNULL}(x) <- value

\S4method{round}{BPCellsMatrix}(x, digits = 0)

\S4method{expm1}{BPCellsMatrix}(x)

\S4method{expm1_slow}{BPCellsMatrix}(x)

\S4method{log1p}{BPCellsMatrix}(x)

\S4method{log1p_slow}{BPCellsMatrix}(x)

\S4method{pmin_by_col}{BPCellsMatrix}(object, values)

\S4method{pmin_by_row}{BPCellsMatrix}(object, values)

\S4method{pmin_scalar}{BPCellsMatrix}(object, value)

\S4method{^}{BPCellsMatrix,ANY}(e1, e2)

\S4method{^}{BPCellsMatrix,numeric}(e1, e2)
}
\arguments{
\item{x, object}{A \code{BPCellsMatrix} object.}

\item{i, j}{Row and Column index.}

\item{...}{Not used currently.}

\item{drop}{Ignored, always be \code{FALSE}.}

\item{value}{\itemize{
\item \verb{dimnames<-}: A list of dimnames or \code{NULL}.
\item \verb{[<-}: A matrix which can be coerced into
\link[Matrix:dgCMatrix-class]{dgCMatrix}.
\item \code{pmin_scalar}: Single positive numeric value.
}}

\item{digits}{Integer indicating the number of decimal places}

\item{values}{A positive atomic numeric.}

\item{e1, e2}{See methods signatures.}
}
\value{
\itemize{
\item \code{t}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \code{[}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \verb{[<-}: A \link{BPCellsMatrix} object.
}

\itemize{
\item \verb{dimnames<-}: A \link[=BPCellsRenameDims]{BPCellsRenameDimsMatrix} object.
}

\itemize{
\item \code{round}: Rounding of matrix Numbers.
}

\itemize{
\item \code{expm1} and \code{expm1_slow}: compute \code{exp(x)-1} of matrix.
}
}
\description{
The \code{BPCellsMatrix} class just inherits from the
\link[DelayedArray:DelayedArray-class]{DelayedMatrix} object. The purpose for
\code{BPCellsMatrix} object is to provide the common methods for all Delayed
BPCells matrix.
}
\section{Slots}{

\describe{
\item{\code{seed}}{A \link[=BPCellsSeed-class]{BPCellsSeed} object.}
}}

\seealso{
\itemize{
\item \link[=BPCells-bind]{bind}: Combine two Objects by Columns or Rows.
\item \link[=BPCells-Multiplication]{\%*\%}: Matrix Multiplication.
\item \link[=BPCells-crossprod]{crossprod}: Matrix Crossproduct.
\item \link[=BPCells-Summarization]{Summarize}: row/col summarization.
\item \link[=BPCells-binarize]{binarize}: Convert matrix elements to zeros and ones.
}
}
