% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BPCellsMatrix.R, R/Class-Dir.R,
%   R/Class-HDF5.R, R/Class-Mem.R
\docType{class}
\name{BPCellsMatrix-class}
\alias{BPCellsMatrix-class}
\alias{BPCellsArray}
\alias{BPCellsMatrix}
\alias{BPCellsArray-class}
\alias{DelayedArray,BPCellsSeed-method}
\alias{matrixClass,BPCellsArray-method}
\alias{DelayedArray,BPCellsDirSeed-method}
\alias{BPCellsDirMatrix-class}
\alias{matrixClass,BPCellsDirArray-method}
\alias{DelayedArray,BPCellsHDF5Seed-method}
\alias{BPCellsHDF5Matrix-class}
\alias{matrixClass,BPCellsHDF5Array-method}
\alias{BPCellsMemArray-class}
\alias{DelayedArray,BPCellsMemSeed-method}
\alias{BPCellsMemMatrix-class}
\alias{matrixClass,BPCellsMemArray-method}
\title{Base Class for Delayed BPCells matrix}
\usage{
BPCellsArray(x)

BPCellsMatrix(x)

\S4method{DelayedArray}{BPCellsSeed}(seed)

\S4method{matrixClass}{BPCellsArray}(x)

\S4method{DelayedArray}{BPCellsDirSeed}(seed)

\S4method{matrixClass}{BPCellsDirArray}(x)

\S4method{DelayedArray}{BPCellsHDF5Seed}(seed)

\S4method{matrixClass}{BPCellsHDF5Array}(x)

\S4method{DelayedArray}{BPCellsMemSeed}(seed)

\S4method{matrixClass}{BPCellsMemArray}(x)
}
\arguments{
\item{x}{\itemize{
\item \code{BPCellsArray} and \code{BPCellsMatrix}: Details see \link{BPCellsSeed} for
supported object.
\item \code{matrixClass}: A BPCellsArray object.
}}

\item{seed}{A \code{BPCellsSeed} object.}
}
\description{
The \code{BPCellsMatrix} class just inherits from the
\link[DelayedArray:DelayedArray-class]{DelayedMatrix} class.
}
\section{Slots}{

\describe{
\item{\code{seed}}{A \link[=BPCellsSeed-class]{BPCellsSeed} object.}
}}

