% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Multiply.R
\docType{class}
\name{BPCellsMultiply}
\alias{BPCellsMultiply}
\alias{BPCellsMultiplyArray-class}
\alias{DelayedArray,BPCellsMultiplySeed-method}
\alias{BPCellsMultiplyArray}
\alias{BPCellsMultiplyMatrix-class}
\alias{matrixClass,BPCellsMultiplyArray-method}
\title{Delayed BPCells MatrixMultiply}
\usage{
\S4method{DelayedArray}{BPCellsMultiplySeed}(seed)

BPCellsMultiplyArray(x)

\S4method{matrixClass}{BPCellsMultiplyArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsMultiplySeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsMatrixMultiplyArray}: A \code{MatrixMultiply} object.
\item \code{matrixClass}: A \code{BPCellsMatrixMultiplyArray} object.
}}
}
\description{
The \code{BPCellsMultiplyArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixMultiply}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use \code{\%*\%}
or \code{crossprod} methods of other \link{BPCellsMatrix} objects.
}
\seealso{
\link{BPCellsSeed}
}
