% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-RankTransform.R
\docType{class}
\name{BPCellsRankTransform}
\alias{BPCellsRankTransform}
\alias{BPCellsRankTransformSeed}
\alias{BPCellsRankTransformArray-class}
\alias{DelayedArray,BPCellsRankTransformSeed-method}
\alias{BPCellsRankTransformArray}
\alias{BPCellsRankTransformMatrix-class}
\alias{matrixClass,BPCellsRankTransformArray-method}
\alias{path,BPCellsRankTransformSeed-method}
\alias{[,BPCellsRankTransformSeed,ANY,ANY,ANY-method}
\title{Delayed BPCells MatrixRankTransform}
\usage{
BPCellsRankTransformSeed(x)

\S4method{DelayedArray}{BPCellsRankTransformSeed}(seed)

BPCellsRankTransformArray(x)

\S4method{matrixClass}{BPCellsRankTransformArray}(x)

\S4method{path}{BPCellsRankTransformSeed}(object)

\S4method{[}{BPCellsRankTransformSeed,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{MatrixRankTransform} object.}

\item{seed}{A \code{BPCellsRankTransformSeed} object.}

\item{object}{A \code{BPCellsRankTransformSeed} object.}

\item{i, j}{Row and Column index.}

\item{...}{Ignored, Not used curretly.}

\item{drop}{Ignored, always be \code{FALSE}.}
}
\description{
The \code{BPCellsRankTransformArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixRankTransform}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link{rank_transform} function to create a \code{BPCellsRankTransformMatrix}.
}
\seealso{
\link[BPCells:convert_matrix_type]{convert_matrix_type}
}
