% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-RenameDims.R
\docType{class}
\name{BPCellsRenameDims}
\alias{BPCellsRenameDims}
\alias{BPCellsRenameDimsArray-class}
\alias{DelayedArray,BPCellsRenameDimsSeed-method}
\alias{BPCellsRenameDimsArray}
\alias{BPCellsRenameDimsMatrix-class}
\alias{matrixClass,BPCellsRenameDimsArray-method}
\title{Delayed BPCells RenameDims}
\usage{
\S4method{DelayedArray}{BPCellsRenameDimsSeed}(seed)

BPCellsRenameDimsArray(x)

\S4method{matrixClass}{BPCellsRenameDimsArray}(x)
}
\arguments{
\item{seed}{A \code{BPCellsRenameDimsSeed} object.}

\item{x}{For Specific functions:
\itemize{
\item \code{BPCellsRenameDimsArray}: A \code{RenameDims} object.
\item \code{matrixClass}: A \code{BPCellsRenameDimsArray} object.
}}
}
\description{
The \code{BPCellsRenameDimsArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{RenameDims}
object in BPCells.
}
\note{
Usually, you shouldn't use this class directly, instead, you should use
\link{dimnames<-} to create a \code{BPCellsRenameDimsSeed} object.
}
\seealso{
\link{BPCellsSeed}
}
