% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BPCellsSeed.R
\name{BPCellsSeed-class}
\alias{BPCellsSeed-class}
\alias{IterableMatrix}
\alias{type,IterableMatrix-method}
\alias{as.array.IterableMatrix}
\alias{as.array,IterableMatrix-method}
\alias{extract_array,IterableMatrix-method}
\alias{OLD_extract_sparse_array,IterableMatrix-method}
\alias{extract_sparse_array,IterableMatrix-method}
\alias{chunkdim,IterableMatrix-method}
\title{Seed Contract methods for \code{IterableMatrix}}
\usage{
\S4method{type}{IterableMatrix}(x)

\method{as.array}{IterableMatrix}(x, drop = FALSE)

\S4method{as.array}{IterableMatrix}(x, drop = FALSE)

\S4method{extract_array}{IterableMatrix}(x, index)

\S4method{OLD_extract_sparse_array}{IterableMatrix}(x, index)

\S4method{extract_sparse_array}{IterableMatrix}(x, index)

\S4method{chunkdim}{IterableMatrix}(x)
}
\arguments{
\item{x, object}{A \code{IterableMatrix} object.}

\item{index}{
    An unnamed list of integer vectors, one per dimension in \code{x}.
    Each vector is called a \emph{subscript} and can only contain
    positive integers that are valid 1-based indices along the corresponding
    dimension in \code{x}.

    Empty or missing subscripts are allowed. They must be represented
    by list elements set to \code{integer(0)} or \code{NULL}, respectively.

    The subscripts cannot contain NAs or non-positive values.

    Individual subscripts are allowed to contain duplicated indices.
  }

\item{value}{\itemize{
\item \verb{dimnames<-}: A list of dimnames or \code{NULL}.
\item \verb{[<-}: A matrix which can be coerced into
\link[Matrix:dgCMatrix-class]{dgCMatrix}.
\item \code{pmin_scalar}: Single positive numeric value.
}}

\item{...}{Not used currently.}
}
\value{
\itemize{
\item \code{type}: A string, indicates the storage type. For all BPCells matrix type
of \code{float} and \code{double}, always return \code{double} since R cannot
differentiate 32-bit and 64-bit real number. See
\link[=convert_mode]{storage_mode}.
}

\itemize{
\item \code{extract_array}: A dense matrix.
}

\itemize{
\item \code{OLD_extract_sparse_array}: A
\link[DelayedArray:SparseArraySeed-class]{SparseArraySeed} object.
}

\itemize{
\item \code{extract_sparse_array}: A
\link[SparseArray:SVT_SparseArray-class]{SparseArray} object.
}

\itemize{
\item \code{chunkdim}: the chunk dimensions in an integer vector parallel to \code{dim(x)}.
}
}
\description{
Seed Contract methods for \code{IterableMatrix}
}
\seealso{
\itemize{
\item \link[=BPCells-bind]{bind}: Combine two Objects by Columns or Rows.
\item \link[=BPCells-Multiplication]{\%*\%}: Matrix Multiplication.
\item \link[=BPCells-crossprod]{crossprod}: Matrix Crossproduct.
\item \link[=BPCells-Summarization]{summarization}: row/col summarization.
\item \link[=BPCells-Arithmetic]{Arithmetic}: Binary Arithmetic operators.
\item \link[=BPCells-binarize]{binarize}: Convert matrix elements to zeros and ones.
}
}
