% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPCellsSeed.R, R/Class-BindMatrix.R,
%   R/Class-Convert.R, R/Class-Mask.R, R/Class-Multiply.R
\name{BPCellsSeed}
\alias{BPCellsSeed}
\alias{BPCellsSeed,IterableMatrix-method}
\alias{BPCellsSeed,BPCellsDelayedOp-method}
\alias{BPCellsSeed,matrix-method}
\alias{BPCellsSeed,dgCMatrix-method}
\alias{BPCellsSeed,ANY-method}
\alias{BPCellsSeed,ColBindMatrices-method}
\alias{BPCellsSeed,RowBindMatrices-method}
\alias{summary.BPCellsDelayedConvert}
\alias{BPCellsSeed,MatrixMask-method}
\alias{BPCellsSeed,MatrixMultiply-method}
\title{Build \code{BPCellsSeed} object}
\usage{
BPCellsSeed(x)

\S4method{BPCellsSeed}{IterableMatrix}(x)

\S4method{BPCellsSeed}{BPCellsDelayedOp}(x)

\S4method{BPCellsSeed}{matrix}(x)

\S4method{BPCellsSeed}{dgCMatrix}(x)

\S4method{BPCellsSeed}{ANY}(x)

\S4method{BPCellsSeed}{ColBindMatrices}(x)

\S4method{BPCellsSeed}{RowBindMatrices}(x)

summary.BPCellsDelayedConvert(object)

\S4method{BPCellsSeed}{MatrixMask}(x)

\S4method{BPCellsSeed}{MatrixMultiply}(x)
}
\arguments{
\item{x}{A \code{IterableMatrix} object from \code{BPCells}, a matrix-like object which
can be coerced into dgCMatrix, or a \link[=BPCellsSeed-class]{BPCellsSeed} object.}
}
\value{
A \link[=BPCellsSeed-class]{BPCellsSeed} object.
}
\description{
Build \code{BPCellsSeed} object
}
\seealso{
\linkS4class{BPCellsSeed}
}
