% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BPCellsSeed.R
\docType{class}
\name{BPCellsSeed}
\alias{BPCellsSeed}
\alias{BPCellsSeed,BPCellsSeed-method}
\alias{BPCellsSeed,MatrixSubset-method}
\alias{BPCellsSeed,ConvertMatrixType-method}
\alias{BPCellsSeed,MatrixMultiply-method}
\alias{BPCellsSeed,RenameDims-method}
\alias{BPCellsSeed,ColBindMatrices-method}
\alias{BPCellsSeed,RowBindMatrices-method}
\alias{BPCellsSeed,MatrixDir-method}
\alias{show,BPCellsSeed-method}
\alias{type,BPCellsSeed-method}
\alias{is_sparse,BPCellsSeed-method}
\alias{extract_array,BPCellsSeed-method}
\alias{extract_sparse_array,BPCellsSeed-method}
\alias{t,BPCellsSeed-method}
\alias{[<-,BPCellsSeed,ANY,ANY,ANY-method}
\alias{matrix_multi}
\alias{matrix_multi,BPCellsSeed,BPCellsSeed-method}
\alias{matrix_multi,BPCellsSeed,dgCMatrix-method}
\alias{matrix_multi,dgCMatrix,BPCellsSeed-method}
\alias{matrix_multi,BPCellsSeed,ANY-method}
\alias{matrix_multi,ANY,BPCellsSeed-method}
\alias{\%*\%,BPCellsSeed,BPCellsSeed-method}
\alias{\%*\%,BPCellsSeed,dgCMatrix-method}
\alias{\%*\%,dgCMatrix,BPCellsSeed-method}
\alias{\%*\%,BPCellsSeed,ANY-method}
\alias{\%*\%,ANY,BPCellsSeed-method}
\alias{crossprod,BPCellsSeed,BPCellsSeed-method}
\alias{crossprod,BPCellsSeed,dgCMatrix-method}
\alias{crossprod,dgCMatrix,BPCellsSeed-method}
\alias{crossprod,BPCellsSeed,matrix-method}
\alias{crossprod,matrix,BPCellsSeed-method}
\alias{crossprod,BPCellsSeed,ANY-method}
\alias{crossprod,ANY,BPCellsSeed-method}
\title{Low-level Base Class for Delayed BPCells matrix}
\usage{
BPCellsSeed(x, ...)

\S4method{BPCellsSeed}{BPCellsSeed}(x, ...)

\S4method{BPCellsSeed}{MatrixSubset}(x, ...)

\S4method{BPCellsSeed}{ConvertMatrixType}(x, ...)

\S4method{BPCellsSeed}{MatrixMultiply}(x, ...)

\S4method{BPCellsSeed}{RenameDims}(x, ...)

\S4method{BPCellsSeed}{ColBindMatrices}(x, ...)

\S4method{BPCellsSeed}{RowBindMatrices}(x, ...)

\S4method{BPCellsSeed}{MatrixDir}(x, ...)

\S4method{show}{BPCellsSeed}(object)

\S4method{type}{BPCellsSeed}(x)

\S4method{is_sparse}{BPCellsSeed}(x)

\S4method{extract_array}{BPCellsSeed}(x, index)

\S4method{extract_sparse_array}{BPCellsSeed}(x, index)

\S4method{t}{BPCellsSeed}(x)

\S4method{[}{BPCellsSeed,ANY,ANY,ANY}(x, i, j, ...) <- value

matrix_multi(x, y, ...)

\S4method{matrix_multi}{BPCellsSeed,BPCellsSeed}(x, y, ...)

\S4method{matrix_multi}{BPCellsSeed,dgCMatrix}(x, y)

\S4method{matrix_multi}{dgCMatrix,BPCellsSeed}(x, y)

\S4method{matrix_multi}{BPCellsSeed,ANY}(x, y)

\S4method{matrix_multi}{ANY,BPCellsSeed}(x, y)

\S4method{\%*\%}{BPCellsSeed,BPCellsSeed}(x, y)

\S4method{\%*\%}{BPCellsSeed,dgCMatrix}(x, y)

\S4method{\%*\%}{dgCMatrix,BPCellsSeed}(x, y)

\S4method{\%*\%}{BPCellsSeed,ANY}(x, y)

\S4method{\%*\%}{ANY,BPCellsSeed}(x, y)

\S4method{crossprod}{BPCellsSeed,BPCellsSeed}(x, y)

\S4method{crossprod}{BPCellsSeed,dgCMatrix}(x, y)

\S4method{crossprod}{dgCMatrix,BPCellsSeed}(x, y)

\S4method{crossprod}{BPCellsSeed,matrix}(x, y)

\S4method{crossprod}{matrix,BPCellsSeed}(x, y)

\S4method{crossprod}{BPCellsSeed,ANY}(x, y)

\S4method{crossprod}{ANY,BPCellsSeed}(x, y)
}
\arguments{
\item{x}{A \link{BPCellsSeed} object. For some functions:
\itemize{
\item \code{BPCellsSeed}: a \link{BPCellsSeed} object, or other BPCells matrix object.
\item \code{matrix_multi}, \code{\%*\%}, and \code{crossprod}: A \link{BPCellsSeed} object or
matrx-like object which can be coerced into a
\link[Matrix:dgCMatrix-class]{dgCMatrix}.
}}

\item{...}{Additional arguments passed to specific methods.
For some functions:
\itemize{
\item \code{matrix_multi}: other arguments passed to
\link[BPCells:transpose_storage_order]{transpose_storage_order}.
}}

\item{object}{A \code{BPCellsSeed} object.}

\item{index}{
    An unnamed list of integer vectors, one per dimension in \code{x}.
    Each vector is called a \emph{subscript} and can only contain
    positive integers that are valid 1-based indices along the corresponding
    dimension in \code{x}.

    Empty or missing subscripts are allowed. They must be represented
    by list elements set to \code{integer(0)} or \code{NULL}, respectively.

    The subscripts cannot contain NAs or non-positive values.

    Individual subscripts are allowed to contain duplicated indices.
  }

\item{i, j}{Row and Column index.}

\item{y}{A \link{BPCellsSeed} object or matrx-like object which can be coerced
into a \link[Matrix:dgCMatrix-class]{dgCMatrix}.}
}
\value{
\itemize{
\item \code{matrix_multi}: matrix multiplication, a \link{BPCellsSeed} object
}

\itemize{
\item \code{x \%*\% y}: matrix multiplication, a \link{BPCellsSeed} object
}

\itemize{
\item \code{crossprod(x, y)}: Matrix Crossproduct, a \link{BPCellsSeed} object
}
}
\description{
The \code{BPCellsSeed} class just inherits from the \code{IterableMatrix} object in
BPCells package. The purpose for \code{BPCellsSeed} object is to provide the
common methods for all low-level BPCells seed objects.
}
