% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPCellsSeed.R
\name{BPCellsSeed}
\alias{BPCellsSeed}
\alias{BPCellsSeed,IterableMatrix-method}
\alias{BPCellsSeed,matrix-method}
\alias{BPCellsSeed,dgCMatrix-method}
\alias{BPCellsSeed,ANY-method}
\title{Transform into \code{IterableMatrix}}
\usage{
BPCellsSeed(x)

\S4method{BPCellsSeed}{IterableMatrix}(x)

\S4method{BPCellsSeed}{matrix}(x)

\S4method{BPCellsSeed}{dgCMatrix}(x)

\S4method{BPCellsSeed}{ANY}(x)
}
\arguments{
\item{x}{A \code{IterableMatrix} object from \code{BPCells} or a matrix-like object which can be coerced into \link[Matrix:dgCMatrix-class]{dgCMatrix} object.}
}
\value{
A \code{IterableMatrix} object.
}
\description{
Transform into \code{IterableMatrix}
}
\seealso{
\itemize{
\item \link{BPCellsSeed}
\item \link{BPCellsDir-IO}
\item \link{BPCellsHDF5-IO}
\item \link{BPCellsMem-IO}
\item \link{BPCells10xHDF5-IO}
\item \link{BPCellsAnnHDF5-IO}
}
}
