% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-BPCellsSeed.R, R/Class-BindMatrix.R,
%   R/Class-Convert.R, R/Class-Dir.R, R/Class-HDF5.R, R/Class-Mask.R,
%   R/Class-Mem.R, R/Class-Multiply.R, R/Class-RankTransform.R,
%   R/Class-Transformed.R, R/Method-subset.R
\docType{class}
\name{BPCellsSeed}
\alias{BPCellsSeed}
\alias{BPCellsSeed-class}
\alias{BPCellsSeed,BPCellsSeed-method}
\alias{BPCellsSeed,Iterable_dgCMatrix_wrapper-method}
\alias{BPCellsSeed,RenameDims-method}
\alias{BPCellsSeed,ColBindMatrices-method}
\alias{BPCellsSeed,RowBindMatrices-method}
\alias{BPCellsSeed,ConvertMatrixType-method}
\alias{BPCellsSeed,MatrixDir-method}
\alias{BPCellsDirArray-class}
\alias{BPCellsSeed,MatrixH5-method}
\alias{BPCellsHDF5Array-class}
\alias{BPCellsSeed,MatrixMask-method}
\alias{BPCellsSeed,PackedMatrixMemBase-method}
\alias{BPCellsSeed,UnpackedMatrixMemBase-method}
\alias{BPCellsSeed,MatrixMultiply-method}
\alias{BPCellsSeed,MatrixRankTransform-method}
\alias{BPCellsSeed,TransformedMatrix-method}
\alias{BPCellsSeed,MatrixSubset-method}
\title{Low-level Base Class for Delayed BPCells matrix}
\usage{
BPCellsSeed(x)

\S4method{BPCellsSeed}{BPCellsSeed}(x)

\S4method{BPCellsSeed}{Iterable_dgCMatrix_wrapper}(x)

\S4method{BPCellsSeed}{RenameDims}(x)

\S4method{BPCellsSeed}{ColBindMatrices}(x)

\S4method{BPCellsSeed}{RowBindMatrices}(x)

\S4method{BPCellsSeed}{ConvertMatrixType}(x)

\S4method{BPCellsSeed}{MatrixDir}(x)

\S4method{BPCellsSeed}{MatrixH5}(x)

\S4method{BPCellsSeed}{MatrixMask}(x)

\S4method{BPCellsSeed}{PackedMatrixMemBase}(x)

\S4method{BPCellsSeed}{UnpackedMatrixMemBase}(x)

\S4method{BPCellsSeed}{MatrixMultiply}(x)

\S4method{BPCellsSeed}{MatrixRankTransform}(x)

\S4method{BPCellsSeed}{TransformedMatrix}(x)

\S4method{BPCellsSeed}{MatrixSubset}(x)
}
\arguments{
\item{x}{A \code{BPCellsSeed} object or a \code{IterableMatrix} object from BPCells,
details see the method signature.}
}
\description{
The \code{BPCellsSeed} class just inherits from the \code{IterableMatrix} class in
BPCells package. The purpose for \code{BPCellsSeed} class is to provide the common
methods for all low-level BPCells seed classes.
}
