% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPCellsSeed.R
\name{BPCellsSeed}
\alias{BPCellsSeed}
\alias{BPCellsSeed,IterableMatrix-method}
\alias{BPCellsSeed,BPCellsArray-method}
\alias{BPCellsSeed,BPCellsMatrix-method}
\alias{BPCellsSeed,BPCellsDelayedOp-method}
\alias{BPCellsSeed,matrix-method}
\alias{BPCellsSeed,dgCMatrix-method}
\alias{BPCellsSeed,ANY-method}
\title{Build seed object for \code{BPCellsMatrix}}
\usage{
BPCellsSeed(x)

\S4method{BPCellsSeed}{IterableMatrix}(x)

\S4method{BPCellsSeed}{BPCellsArray}(x)

\S4method{BPCellsSeed}{BPCellsMatrix}(x)

\S4method{BPCellsSeed}{BPCellsDelayedOp}(x)

\S4method{BPCellsSeed}{matrix}(x)

\S4method{BPCellsSeed}{dgCMatrix}(x)

\S4method{BPCellsSeed}{ANY}(x)
}
\arguments{
\item{x}{A \code{IterableMatrix} object from \code{BPCells} or a matrix-like object
which can be coerced into dgCMatrix object.}
}
\value{
A \code{IterableMatrix} object.
}
\description{
Build seed object for \code{BPCellsMatrix}
}
\seealso{
\link[=BPCellsSeed-class]{BPCellsSeed} contract
}
