% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPCellsSeed.R
\name{BPCellsSeed}
\alias{BPCellsSeed}
\alias{BPCellsMatrix}
\alias{BPCellsSeed,BPCellsSeed-method}
\alias{BPCellsSeed,MatrixSubset-method}
\alias{BPCellsSeed,ConvertMatrixType-method}
\alias{BPCellsSeed,MatrixMultiply-method}
\alias{BPCellsSeed,RenameDims-method}
\alias{BPCellsSeed,TransformedMatrix-method}
\alias{BPCellsSeed,TransformLog1p-method}
\alias{BPCellsSeed,TransformLog1pSlow-method}
\alias{BPCellsSeed,TransformExpm1-method}
\alias{BPCellsSeed,TransformExpm1Slow-method}
\alias{BPCellsSeed,TransformSquare-method}
\alias{BPCellsSeed,TransformPow-method}
\alias{BPCellsSeed,TransformPowSlow-method}
\alias{BPCellsSeed,TransformMin-method}
\alias{BPCellsSeed,TransformMinByRow-method}
\alias{BPCellsSeed,TransformMinByCol-method}
\alias{BPCellsSeed,TransformBinarize-method}
\alias{BPCellsSeed,TransformRound-method}
\alias{BPCellsSeed,MatrixRankTransform-method}
\alias{BPCellsSeed,MatrixMask-method}
\alias{BPCellsSeed,ColBindMatrices-method}
\alias{BPCellsSeed,RowBindMatrices-method}
\alias{BPCellsSeed,MatrixDir-method}
\title{Create a Low-level Base Class for Delayed BPCells matrix}
\usage{
BPCellsMatrix(x, ...)

BPCellsSeed(x, ...)

\S4method{BPCellsSeed}{BPCellsSeed}(x, ...)

\S4method{BPCellsSeed}{MatrixSubset}(x, ...)

\S4method{BPCellsSeed}{ConvertMatrixType}(x, ...)

\S4method{BPCellsSeed}{MatrixMultiply}(x, ...)

\S4method{BPCellsSeed}{RenameDims}(x, ...)

\S4method{BPCellsSeed}{TransformedMatrix}(x, ...)

\S4method{BPCellsSeed}{TransformLog1p}(x, ...)

\S4method{BPCellsSeed}{TransformLog1pSlow}(x, ...)

\S4method{BPCellsSeed}{TransformExpm1}(x, ...)

\S4method{BPCellsSeed}{TransformExpm1Slow}(x, ...)

\S4method{BPCellsSeed}{TransformSquare}(x, ...)

\S4method{BPCellsSeed}{TransformPow}(x, ...)

\S4method{BPCellsSeed}{TransformPowSlow}(x, ...)

\S4method{BPCellsSeed}{TransformMin}(x, ...)

\S4method{BPCellsSeed}{TransformMinByRow}(x, ...)

\S4method{BPCellsSeed}{TransformMinByCol}(x, ...)

\S4method{BPCellsSeed}{TransformBinarize}(x, ...)

\S4method{BPCellsSeed}{TransformRound}(x, ...)

\S4method{BPCellsSeed}{MatrixRankTransform}(x, ...)

\S4method{BPCellsSeed}{MatrixMask}(x, ...)

\S4method{BPCellsSeed}{ColBindMatrices}(x, ...)

\S4method{BPCellsSeed}{RowBindMatrices}(x, ...)

\S4method{BPCellsSeed}{MatrixDir}(x, buffer_size = 8192L)
}
\arguments{
\item{x}{A BPCells IterableMatrix object.}

\item{...}{\itemize{
\item \code{BPCellsSeed}: Additional arguments passed to specific methods.
\item \code{BPCellsMatrix}: Additional arguments passed to \code{BPCellsSeed}.
}}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}
}
\value{
\itemize{
\item \code{BPCellsSeed}: A \link{BPCellsSeed} object.
\item \code{BPCellsMatrix}: A \link{BPCellsMatrix} object.
}
}
\description{
Create a Low-level Base Class for Delayed BPCells matrix
}
