% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Subset.R
\docType{class}
\name{BPCellsSubset}
\alias{BPCellsSubset}
\alias{BPCellsSubsetSeed}
\alias{BPCellsSubsetArray-class}
\alias{DelayedArray,BPCellsSubsetSeed-method}
\alias{BPCellsSubsetArray}
\alias{BPCellsSubsetMatrix-class}
\alias{matrixClass,BPCellsSubsetArray-method}
\alias{path,BPCellsSubsetSeed-method}
\title{Delayed BPCells MatrixSubset}
\usage{
BPCellsSubsetSeed(x)

\S4method{DelayedArray}{BPCellsSubsetSeed}(seed)

BPCellsSubsetArray(x)

\S4method{matrixClass}{BPCellsSubsetArray}(x)

\S4method{path}{BPCellsSubsetSeed}(object)
}
\arguments{
\item{x}{A \code{MatrixSubset} or \code{BPCellsSubsetSeed} object.}

\item{seed}{A \code{BPCellsSubsetSeed} object.}

\item{object}{A \code{BPCellsSubsetSeed} object.}
}
\description{
The \code{BPCellsSubsetArray} class provides a
\link[DelayedArray:DelayedArray-class]{DelayedArray} backend for \code{MatrixSubset} object
in BPCells. Usually, you shouldn't use this class directly, instead, you
should use \code{[} (extract methods) of other BPCellsArray objects.
}
