% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-Convert.R
\name{convert_mode}
\alias{convert_mode}
\alias{convert_mode,BPCellsMatrix-method}
\alias{storage_mode}
\alias{storage_mode,BPCellsMatrix-method}
\alias{storage_mode,BPCellsDelayedOp-method}
\alias{storage_mode,IterableMatrix-method}
\alias{storage_mode,matrix-method}
\title{Convert the storage mode of a BPCellsArray object}
\usage{
convert_mode(object, ...)

\S4method{convert_mode}{BPCellsMatrix}(object, mode)

storage_mode(object)

\S4method{storage_mode}{BPCellsMatrix}(object)

\S4method{storage_mode}{BPCellsDelayedOp}(object)

\S4method{storage_mode}{IterableMatrix}(object)

\S4method{storage_mode}{matrix}(object)
}
\arguments{
\item{object}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{...}{Additional parameters passed into specific methods.}

\item{mode}{Storage mode of BPCells matrix, one of \code{uint32_t} (unsigned
32-bit integer), \code{float} (32-bit real number), or \code{double} (64-bit real
number). R cannot differentiate 32-bit and 64-bit real number, so
\link[=BPCellsSeed-class]{type} method always return "double" for both \code{float}
and \code{double} mode.}
}
\value{
\itemize{
\item \code{convert_mode}: A \link[=BPCellsMatrix-class]{BPCellsMatrix} object with storage
mode converted into the specified.
}

\itemize{
\item \code{storage_mode}: A string indicates the storage mode.
}
}
\description{
Convert the storage mode of a BPCellsArray object
}
\seealso{
\link[BPCells:convert_matrix_type]{convert_matrix_type}
}
