% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Convert.R
\name{convert_type}
\alias{convert_type}
\alias{convert_type,BPCellsSeed-method}
\alias{convert_type,BPCellsMatrix-method}
\title{Convert the storage type of a BPCellsArray object}
\usage{
convert_type(object, ...)

\S4method{convert_type}{BPCellsSeed}(object, type)

\S4method{convert_type}{BPCellsMatrix}(object, type)
}
\arguments{
\item{object}{A \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{...}{Additional parameters passed into specific methods.}

\item{type}{Storage mode of BPCells matrix, one of \code{uint32_t} (\code{integer})
(unsigned 32-bit integer), \code{float} (\code{numeric} or \verb{32bit_numeric}) (32-bit
real number), or \code{double} (\verb{64bit_numeric}) (64-bit real number). R cannot
differentiate 32-bit and 64-bit real number, here, we use "double" to indicte
64-bit real number and "numeric" to indicate 32-bit real number.}
}
\value{
A \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
Convert the storage type of a BPCellsArray object
}
\seealso{
\link[BPCells:convert_matrix_type]{convert_matrix_type}
}
