% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Convert.R, R/Class-Mask.R,
%   R/Class-RankTransform.R, R/Class-dgCMatrix.R, R/Method-binarize.R,
%   R/Method-bind.R, R/Method-internal.R, R/Transform-min_by_row.R
\name{convert_type,ANY-method}
\alias{convert_type,ANY-method}
\alias{mask_matrix,ANY,ANY-method}
\alias{rank_transform,ANY-method}
\alias{DelayedArray,BPCellsdgCMatrixSeed-method}
\alias{matrixClass,BPCellsdgCMatrixArray-method}
\alias{<,BPCellsMatrix,numeric-method}
\alias{>,numeric,BPCellsMatrix-method}
\alias{<=,BPCellsMatrix,numeric-method}
\alias{>=,numeric,BPCellsMatrix-method}
\alias{<,BPCellsSeed,numeric-method}
\alias{>,numeric,BPCellsSeed-method}
\alias{<=,BPCellsSeed,numeric-method}
\alias{>=,numeric,BPCellsSeed-method}
\alias{rbind2,ANY,BPCellsMatrix-method}
\alias{rbind2,BPCellsMatrix,ANY-method}
\alias{cbind2,ANY,BPCellsMatrix-method}
\alias{cbind2,BPCellsMatrix,ANY-method}
\alias{rbind2,ANY,BPCellsSeed-method}
\alias{rbind2,BPCellsSeed,ANY-method}
\alias{cbind2,ANY,BPCellsSeed-method}
\alias{cbind2,BPCellsSeed,ANY-method}
\alias{internal-methods}
\alias{matrixClass,BPCellsTransformMinByRowArray-method}
\title{House of internal methods}
\usage{
\S4method{convert_type}{ANY}(object, type)

\S4method{mask_matrix}{ANY,ANY}(object, mask, invert = FALSE)

\S4method{rank_transform}{ANY}(object, axis)

\S4method{DelayedArray}{BPCellsdgCMatrixSeed}(seed)

\S4method{matrixClass}{BPCellsdgCMatrixArray}(x)

\S4method{<}{BPCellsMatrix,numeric}(e1, e2)

\S4method{>}{numeric,BPCellsMatrix}(e1, e2)

\S4method{<=}{BPCellsMatrix,numeric}(e1, e2)

\S4method{>=}{numeric,BPCellsMatrix}(e1, e2)

\S4method{<}{BPCellsSeed,numeric}(e1, e2)

\S4method{>}{numeric,BPCellsSeed}(e1, e2)

\S4method{<=}{BPCellsSeed,numeric}(e1, e2)

\S4method{>=}{numeric,BPCellsSeed}(e1, e2)

\S4method{rbind2}{ANY,BPCellsMatrix}(x, y, ...)

\S4method{rbind2}{BPCellsMatrix,ANY}(x, y, ...)

\S4method{cbind2}{ANY,BPCellsMatrix}(x, y, ...)

\S4method{cbind2}{BPCellsMatrix,ANY}(x, y, ...)

\S4method{rbind2}{ANY,BPCellsSeed}(x, y, ...)

\S4method{rbind2}{BPCellsSeed,ANY}(x, y, ...)

\S4method{cbind2}{ANY,BPCellsSeed}(x, y, ...)

\S4method{cbind2}{BPCellsSeed,ANY}(x, y, ...)

\S4method{matrixClass}{BPCellsTransformMinByRowArray}(x)
}
\arguments{
\item{object}{A \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{type}{Storage mode of BPCells matrix, one of \code{uint32_t} (\code{integer})
(unsigned 32-bit integer), \code{float} (\code{numeric} or \verb{32bit_numeric}) (32-bit
real number), or \code{double} (\verb{64bit_numeric}) (64-bit real number). R cannot
differentiate 32-bit and 64-bit real number, here, we use "double" to indicte
64-bit real number and "numeric" to indicate 32-bit real number.}

\item{mask}{Mask matrix, A \link{BPCellsSeed} or \link{BPCellsMatrix} object.
Additionally, a matrix-like object which can be coerced into
\link[Matrix:dgCMatrix-class]{dgCMatrix}.}

\item{invert}{A bool, indicates whether revert the mask.}

\item{axis}{Axis to rank values within. "col" to rank values within each
column, and "row" to rank values within each row. If \code{NULL}, will use the
storage order of \code{object} (see \link[BPCells:IterableMatrix-methods]{storage_order}).
If axis value is different from the storage order of \code{object},
\link[BPCells:transpose_storage_order]{transpose_storage_order} will be used
to transpose the underlying storage order.}

\item{seed}{A \link[=BPCellsSeed-class]{BPCellsSeed} object.}

\item{x, y}{A \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{...}{Additional parameters passed into specific methods.}
}
\description{
Following methods are used by package internal, for messages purpose, usually
the method for \code{ANY} object.
}
