% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Mask.R
\name{mask_matrix}
\alias{mask_matrix}
\alias{mask_matrix,BPCellsSeed,BPCellsSeed-method}
\alias{mask_matrix,BPCellsSeed,dgCMatrix-method}
\alias{mask_matrix,BPCellsSeed,BPCellsMatrix-method}
\alias{mask_matrix,BPCellsSeed,ANY-method}
\alias{mask_matrix,BPCellsMatrix,BPCellsSeed-method}
\alias{mask_matrix,BPCellsMatrix,dgCMatrix-method}
\alias{mask_matrix,BPCellsMatrix,BPCellsMatrix-method}
\alias{mask_matrix,BPCellsMatrix,ANY-method}
\title{Mask matrix entries to zero}
\usage{
mask_matrix(object, mask, ...)

\S4method{mask_matrix}{BPCellsSeed,BPCellsSeed}(object, mask, invert = FALSE)

\S4method{mask_matrix}{BPCellsSeed,dgCMatrix}(object, mask, invert = FALSE)

\S4method{mask_matrix}{BPCellsSeed,BPCellsMatrix}(object, mask, invert = FALSE)

\S4method{mask_matrix}{BPCellsSeed,ANY}(object, mask, invert = FALSE)

\S4method{mask_matrix}{BPCellsMatrix,BPCellsSeed}(object, mask, invert = FALSE)

\S4method{mask_matrix}{BPCellsMatrix,dgCMatrix}(object, mask, invert = FALSE)

\S4method{mask_matrix}{BPCellsMatrix,BPCellsMatrix}(object, mask, invert = FALSE)

\S4method{mask_matrix}{BPCellsMatrix,ANY}(object, mask, invert = FALSE)
}
\arguments{
\item{object}{A \link{BPCellsSeed} or \link{BPCellsMatrix} object.}

\item{mask}{Mask matrix, A \link{BPCellsSeed} or \link{BPCellsMatrix} object.
Additionally, a matrix-like object which can be coerced into
\link[Matrix:dgCMatrix-class]{dgCMatrix}.}

\item{...}{Additional parameters passed into specific methods.}

\item{invert}{A bool, indicates whether revert the mask.}
}
\value{
A \link[=BPCellsMask]{BPCellsMaskSeed} or \link[=BPCellsMask]{BPCellsMatrix}
object.
}
\description{
Set matrix entries to zero given a mask matrix of the same dimensions.
Normally, non-zero values in the mask will set the matrix entry to zero. If
inverted, zero values in the mask matrix will set the matrix entry to zero.
}
\seealso{
\link[BPCells:mask_matrix]{mask_matrix}
}
