% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-pmin.R
\name{pmin2}
\alias{pmin2}
\alias{pmin_by_col}
\alias{pmin_by_col,BPCellsMatrix-method}
\alias{pmin_by_row}
\alias{pmin_by_row,BPCellsMatrix-method}
\alias{pmin_scalar}
\alias{pmin_scalar,BPCellsMatrix-method}
\alias{pmin2,BPCellsMatrix,numeric-method}
\alias{pmax2}
\alias{pmin2,numeric,BPCellsMatrix-method}
\title{Maxima and Minima}
\usage{
pmin_by_col(object, values)

\S4method{pmin_by_col}{BPCellsMatrix}(object, values)

pmin_by_row(object, values)

\S4method{pmin_by_row}{BPCellsMatrix}(object, values)

pmin_scalar(object, value)

\S4method{pmin_scalar}{BPCellsMatrix}(object, value)

\S4method{pmin2}{BPCellsMatrix,numeric}(e1, e2)

\S4method{pmin2}{numeric,BPCellsMatrix}(e1, e2)
}
\arguments{
\item{object}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{values}{An atomic positive numeric.}

\item{value}{A single positive numeric value}

\item{e1, e2}{One of \code{x} or \code{y} must be a \link[=BPCellsMatrix-class]{BPCellsMatrix} object, and the another
must be a scalar or of length \code{nrow(e1)/nrow(e2)}.}
}
\value{
\itemize{
\item \code{pmin_by_col}: Take the minimum with a per-col constant
}

\itemize{
\item \code{pmin_by_row}: Take the minimum with a per-row constant
}

\itemize{
\item \code{pmin_scalar}: Take minumum with a global constant
}
}
\description{
Maxima and Minima
}
\note{
For \code{pmin2} Methods listed here are supported by \code{BPCells}, other
\code{pmin2} methods and \code{pmax2} function will use the methods defined in
\link[=DelayedArray-utils]{DelayedArray}.
}
