% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-RankTransform.R
\name{rank_transform}
\alias{rank_transform}
\alias{rank_transform,BPCellsSeed-method}
\alias{rank_transform,BPCellsMatrix-method}
\alias{rowRanks,BPCellsSeed-method}
\alias{rowRanks}
\alias{colRanks,BPCellsSeed-method}
\alias{colRanks}
\alias{rowRanks,BPCellsMatrix-method}
\alias{colRanks,BPCellsMatrix-method}
\title{Rank-transform a BPCells IterableMatrix matrix}
\usage{
rank_transform(object, ...)

\S4method{rank_transform}{BPCellsSeed}(object, axis = NULL, offset = TRUE, ...)

\S4method{rank_transform}{BPCellsMatrix}(object, axis = NULL, ...)

\S4method{rowRanks}{BPCellsSeed}(
  x,
  rows = NULL,
  cols = NULL,
  ties.method = c("max", "average"),
  ...,
  useNames = NA
)

\S4method{colRanks}{BPCellsSeed}(
  x,
  rows = NULL,
  cols = NULL,
  ties.method = c("max", "average"),
  ...,
  useNames = NA
)

\S4method{rowRanks}{BPCellsMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  ties.method = c("max", "average"),
  ...,
  useNames = NA
)

\S4method{colRanks}{BPCellsMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  ties.method = c("max", "average"),
  ...,
  useNames = NA
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[BPCells:transpose_storage_order]{BPCells::transpose_storage_order}}
  \describe{
    \item{\code{outdir}}{Directory to store the output}
    \item{\code{tmpdir}}{Temporary directory to use for intermediate storage}
    \item{\code{load_bytes}}{The minimum contiguous load size during the merge sort passes}
    \item{\code{sort_bytes}}{The amount of memory to allocate for re-sorting chunks of entries}
  }}

\item{axis}{Axis to rank values within. "col" to rank values within each
column, and "row" to rank values within each row. If \code{NULL}, will use the
storage axis of \code{object} (see \link{storage_axis}).  If \code{axis} specified is
different from the storage axis of \code{object}, \link{transpose_axis} will be
used to transpose the underlying storage order.}

\item{offset}{A bool, whether or not to add offset such that the rank of a 0
value is 0. Default: \code{TRUE}.}

\item{x, object}{A \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{rows, cols, ties.method, useNames}{Ignored currently.}
}
\value{
\itemize{
\item \code{rank_transform}: A \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object depends on the class of \code{object}.
}

\itemize{
\item \code{rowRanks()}: vector of row ranks.
}

\itemize{
\item \code{colRanks()}: vector of column ranks.
}
}
\description{
Rank values are default offset such that the rank of a 0 value is 0. If you
want to get the same result of \code{rowRanks(object, ties.method = "average")}/\code{colRanks(object, ties.method = "average")}, set \code{offset=FALSE}
}
\note{
Ties are always handled by averaging ranks.
}
\seealso{
\link[BPCells:rank_transform]{rank_transform}
}
