% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-RankTransform.R
\name{rank_transform}
\alias{rank_transform}
\alias{rank_transform,BPCellsSeed-method}
\alias{rank_transform,BPCellsMatrix-method}
\title{Rank-transform a BPCells IterableMatrix matrix}
\usage{
rank_transform(object, ...)

\S4method{rank_transform}{BPCellsSeed}(object, axis = NULL, ...)

\S4method{rank_transform}{BPCellsMatrix}(object, axis = NULL, ...)
}
\arguments{
\item{object}{A \code{BPCellsSeed} or \code{BPCellsMatrix} object.}

\item{...}{
  Arguments passed on to \code{\link[BPCells:transpose_storage_order]{BPCells::transpose_storage_order}}
  \describe{
    \item{\code{outdir}}{Directory to store the output}
    \item{\code{tmpdir}}{Temporary directory to use for intermediate storage}
    \item{\code{load_bytes}}{The minimum contiguous load size during the merge sort passes}
    \item{\code{sort_bytes}}{The amount of memory to allocate for re-sorting chunks of entries}
  }}

\item{axis}{Axis to rank values within. "col" to rank values within each
column, and "row" to rank values within each row. If \code{NULL}, will use the
storage order of \code{object} (see \link[BPCells:IterableMatrix-methods]{storage_order}).
If axis value is different from the storage order of \code{object},
\link[BPCells:transpose_storage_order]{transpose_storage_order} will be used
to transpose the underlying storage order.}
}
\value{
A \link[=BPCellsConvert]{BPCellsConverSeed} object or
\link[=BPCellsConvert]{BPCellsConvertMatrix} object.
}
\description{
Rank-transform a BPCells IterableMatrix matrix
}
\seealso{
\link[BPCells:rank_transform]{rank_transform}
}
