% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-BindMatrix.R
\name{set_threads}
\alias{set_threads}
\alias{set_threads,BPCellsMatrix-method}
\title{Set matrix op thread count}
\usage{
set_threads(object, ...)

\S4method{set_threads}{BPCellsMatrix}(object, threads = 0L)
}
\arguments{
\item{object}{A \link{BPCellsMatrix} object with a seed slot of
\link[=BPCellsDelayedOp-class]{BPCellsDelayedAbind} object, usually derived from
\link[=BPCells-bind]{bind} operators.}

\item{...}{Additional arguments to specific methods.}

\item{threads}{Set the number of threads to use for sparse-dense multiply and
\link[BPCells:matrix_stats]{matrix_stats}.}
}
\value{
A \link{BPCellsMatrix} object.
}
\description{
Set number of threads to use for sparse-dense multiply and matrix_stats.
}
