% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showtree.R
\name{showtree}
\alias{showtree}
\alias{seedApply}
\title{Visualize and access the leaves of a tree of delayed operations}
\usage{
showtree(object, show.node.dim = TRUE)

seedApply(x, FUN, ...)
}
\arguments{
\item{show.node.dim}{
    \code{TRUE} or \code{FALSE}. If \code{TRUE} (the default), the nodes
    dimensions and data type are displayed.
  }

\item{x, object}{
    Typically a \link[DelayedArray]{DelayedArray} object but can also be a \link[DelayedArray]{DelayedOp}
    object or a list where each element is a \link[DelayedArray]{DelayedArray} or
    \link[DelayedArray]{DelayedOp} object.
  }

\item{FUN}{
    The function to be applied to each leaf in \code{x}.
  }

\item{...}{
    Optional arguments to \code{FUN} for \code{seedApply()}.

    Additional arguments passed to methods for \code{path()}.
  }
}
\value{
\itemize{
\item \code{showtree}: return the input invisiblely
}

\itemize{
\item \code{seedApply}: A list of length \code{nseed(x)} for \code{seedApply}.
}
}
\description{
\code{showtree} can be used to visualize the tree of delayed operations carried by
a \code{DelayedArray} object.
}
\details{
Use \code{seedApply} to apply a function to the seeds of a \code{DelayedArray} object.
}
