% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showtree.R
\name{showtree}
\alias{showtree}
\alias{nseed,BPCellsMatrix-method}
\alias{seed,BPCellsMatrix-method}
\alias{path,BPCellsMatrix-method}
\alias{path,BPCellsMemSeed-method}
\alias{path,BPCellsdgCMatrixSeed-method}
\alias{path,BPCellsDirSeed-method}
\alias{path,BPCellsHDF5Seed-method}
\alias{seed,BPCellsBasicSeed-method}
\alias{nseed,BPCellsBasicSeed-method}
\alias{path,BPCellsUnaryOpsSeed-method}
\alias{seed,BPCellsUnaryOpsSeed-method}
\alias{nseed,BPCellsUnaryOpsSeed-method}
\alias{path,BPCellsNaryOpsSeed-method}
\alias{seed,BPCellsNaryOpsSeed-method}
\alias{nseed,BPCellsNaryOpsSeed-method}
\title{Visualize and access the leaves of a tree of delayed operations}
\usage{
showtree(object)

\S4method{nseed}{BPCellsMatrix}(x)

\S4method{seed}{BPCellsMatrix}(x)

\S4method{path}{BPCellsMatrix}(object, ...)

\S4method{path}{BPCellsMemSeed}(object, ...)

\S4method{path}{BPCellsdgCMatrixSeed}(object, ...)

\S4method{path}{BPCellsDirSeed}(object, ...)

\S4method{path}{BPCellsHDF5Seed}(object, ...)

\S4method{seed}{BPCellsBasicSeed}(x)

\S4method{nseed}{BPCellsBasicSeed}(x)

\S4method{path}{BPCellsUnaryOpsSeed}(object, ...)

\S4method{seed}{BPCellsUnaryOpsSeed}(x)

\S4method{nseed}{BPCellsUnaryOpsSeed}(x)

\S4method{path}{BPCellsNaryOpsSeed}(object, unlist = TRUE)

\S4method{seed}{BPCellsNaryOpsSeed}(x)

\S4method{nseed}{BPCellsNaryOpsSeed}(x)
}
\arguments{
\item{x, object}{A \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object}

\item{...}{Additional arguments passed to other methods.}

\item{unlist}{A bool, if \code{TRUE}, will unlist the output, used to return a
character vector for \code{path} function, otherwise, \code{path} function will return
a list of path.}
}
\value{
\itemize{
\item \code{showtree}: return the input invisiblely
\item \code{path}: An atomic characters or a list of characters.
\item \code{nseed}: An integer number.
\item \code{seed}: A list of \link[=BPCellsSeed-class]{BPCellsSeed} objects or a single
\link[=BPCellsSeed-class]{BPCellsSeed} object.
}
}
\description{
\code{showtree} can be used to visualize the tree of delayed operations carried by
a \code{BPCellsArray} object.
Use \code{nseed}, \code{seed}, or \code{path} to access the number of seeds, the seed, or
the seed path of a \link[=BPCellsSeed-class]{BPCellsSeed} or
\link[=BPCellsMatrix-class]{BPCellsMatrix} object, respectively.
}
