% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method-axis.R
\name{transpose_axis}
\alias{transpose_axis}
\alias{transpose_axis,BPCellsMatrix-method}
\alias{storage_axis}
\alias{storage_axis,BPCellsMatrix-method}
\alias{storage_axis,BPCellsDelayedOp-method}
\alias{storage_axis,IterableMatrix-method}
\title{Transpose the storage axis for a \code{BPCellsSeed} or \code{BPCellsMatrix} object}
\usage{
transpose_axis(object, ...)

\S4method{transpose_axis}{BPCellsMatrix}(object, ...)

storage_axis(object)

\S4method{storage_axis}{BPCellsMatrix}(object)

\S4method{storage_axis}{BPCellsDelayedOp}(object)

\S4method{storage_axis}{IterableMatrix}(object)
}
\arguments{
\item{object}{A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.}

\item{...}{
  Arguments passed on to \code{\link[BPCells:transpose_storage_order]{BPCells::transpose_storage_order}}
  \describe{
    \item{\code{outdir}}{Directory to store the output}
    \item{\code{tmpdir}}{Temporary directory to use for intermediate storage}
    \item{\code{load_bytes}}{The minimum contiguous load size during the merge sort passes}
    \item{\code{sort_bytes}}{The amount of memory to allocate for re-sorting chunks of entries}
  }}
}
\value{
\itemize{
\item \code{transpose_axis}: A \link{BPCellsMatrix} object with storage axis flipped.
}

\itemize{
\item \code{storage_axis}: A string indicates the storage axis, "row" or "col".
}
}
\description{
Transpose the storage axis for a \code{BPCellsSeed} or \code{BPCellsMatrix} object
}
\details{
This re-sorts the entries of a matrix to change the storage order
from row-major to col-major. For large matrices, this can be slow -- around 2
minutes to transpose a 500k cell RNA-seq matrix The default load_bytes (4MiB)
and sort_bytes (1GiB) parameters allow ~85GB of data to be sorted with two
passes through the data, and ~7.3TB of data to be sorted in three passes
through the data.
}
