% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Mem.R
\name{writeBPCellsMemArray}
\alias{writeBPCellsMemArray}
\alias{writeBPCellsMemArray,ANY-method}
\title{Write a sparce matrices into memory with BPCells format}
\usage{
writeBPCellsMemArray(x, ...)

\S4method{writeBPCellsMemArray}{ANY}(x, compress = TRUE)
}
\arguments{
\item{x}{A \code{IterableMatrix} object from \code{BPCells}, a matrix-like object which
can be coerced into dgCMatrix, or a \code{BPCellsSeed} object.}

\item{...}{For \code{BPCellsMatrix} method: additional parameters passed to \code{ANY}
methods.}

\item{compress}{Whether or not to compress the data.}
}
\value{
A \link[=BPCellsMatrix-class]{BPCellsMatrix} object.
}
\description{
Write a sparce matrices into memory with BPCells format
}
